--- jd_entertainment.js
@@ -1,62 +1,0 @@
-/*
-百变大咖秀
-活动入口：首页搜索-‘百变大咖秀’-底部最右侧按钮
-请手动进入一次活动页面已确保能够自动抽奖
-活动地址：https://lzdz-isv.isvjcloud.com/dingzhi/change/able/activity/3508994?activityId=dz2102100001340207
-
-新手写脚本，难免有bug，能用且用。
-多谢 whyour 大佬 指导
-
-脚本内置了一个给作者任务助力的网络请求，默认开启，如介意请自行关闭。
-助力活动链接： https://h5.m.jd.com/babelDiy/Zeus/4ZK4ZpvoSreRB92RRo8bpJAQNoTq/index.html
-参数 helpAuthor = false
-
-更新地址：https://raw.githubusercontent.com/i-chenzhe/qx/main/jd_entertainment.js
-============Quantumultx===============
-[task_local]
-#百变大咖秀
-10 10,11 * * 2-5 https://raw.githubusercontent.com/i-chenzhe/qx/main/jd_entertainment.js, tag=百变大咖秀,  enabled=true
-================Loon==============
-[Script]
-cron "10 10,11 * * 2-5" script-path=https://raw.githubusercontent.com/i-chenzhe/qx/main/jd_entertainment.js,tag=百变大咖秀
-===============Surge=================
-百变大咖秀 = type=cron,cronexp="10 10,11 * * 2-5",wake-system=1,timeout=3600,script-path=https://raw.githubusercontent.com/i-chenzhe/qx/main/jd_entertainment.js
-============小火箭=========
-百变大咖秀 = type=cron,script-path=https://raw.githubusercontent.com/i-chenzhe/qx/main/jd_entertainment.js, cronexpr="10 10,11 * * 2-5", timeout=3600, enable=true
-*/
-
-const $ = new Env('百变大咖秀');
-const jdCookieNode = $.isNode() ? require('./jdCookie.js') : '';
-const notify = $.isNode() ? require('./sendNotify') : '';
-let cookiesArr = [], cookie = '', originCookie = '', message = '';
-let helpAuthor = true;//为作者助力的开关
-const ACT_ID = 'dz2102100001340207';
-const questionList = [
-  { q: '0c32c1c0e2284331beec6910dc07a948', a: 'A:杨钰莹' },
-  { q: '3576e8750b0b4322977420622aff18a0', a: 'C:王勉' },
-  { q: '43669f0fcf074342934e1422a26d164a', a: 'B:杜海涛' },
-  { q: '7ef574a3f282466ea36cbd4d0b6e1685', a: 'A:王晨艺' },
-  { q: 'ade513a501ce48df9a4816566f24904e', a: 'B:沈凌' },
-  { q: 'a8e35c93ad79460d871f42a10bce68d4', a: 'A:沈梦辰' },
-  { q: 'afd88ed596324ba9a84129112d2fb8b6', a: 'C:锤娜丽莎' },
-  { q: 'c6bacb5e9aac49f0bc4beae144decf80', a: 'B:玲花' },
-  { q: 'e5421dda894e4b43863aad4fd524fd9f', a: 'C:杨迪' }
-]
-if ($.isNode()) {
-  Object.keys(jdCookieNode).forEach((item) => {
-    cookiesArr.push(jdCookieNode[item])
-  })
-  if (process.env.JD_DEBUG && process.env.JD_DEBUG === 'false') console.log = () => { };
-} else {
-  let cookiesData = $.getdata('CookiesJD') || "[]";
-  cookiesData = JSON.parse(cookiesData);
-  cookiesArr = cookiesData.map(item => item.cookie);
-  cookiesArr.reverse();
-  cookiesArr.push(...[$.getdata('CookieJD2'), $.getdata('CookieJD')]);
-  cookiesArr.reverse();
-  cookiesArr = cookiesArr.filter(item => !!item);
-}
-$.log('脚本版本 v0.4\n更新时间:2021-03-19 06:59\n仓库：https://www.github.com/i-chenzhe/qx');
-   var _0xod3='jsjiami.com.v6',_0x516b=[_0xod3,'wqvCnsOadUg=','OsOZwr8=','DcOHwqHChi0=','wqtgw7N4w6U=','ZCvCqMKDKw==','w5ZqABlTwopyHQbCscOcZ8OwwrHCgyIqwo0Nwr7CpcOx','w7NXLMKNw4Q=','XMOLRMKewrZ5w50cwp9Wwrwcw6ALIcOIRg==','IVQFwqlnw4gyw4TDq8Oiw4bClcKhw67CjzA4EcOqQ2vCu8KYOsOTWcKnV3fDkcKQwrYI','b1fDmjHClMOwwpXCt8OvGQ50wqrCoMOsecOgYWI0wppuwqwlw7YhIW5yYw==','L8OywovCmDXCqcOwDMKsw4bCocKywpvCusOGZlDCtcKkbcOUwo8/wp4+esOsXmfDiVwoQsOIwrbDqcOYczAQFiJaw77CtXkIwp3DkyjCqj/CtT7CkDfCnMOEDxsowp7CoMOCw57DnMKtwqRww4nCmW12TMKSwoXDiW3Ct0ZRwrTDgXptwo3ClsK6AQrCjibCvjXDjsKDU1QLNsKKZsKJYcKuwrjDun/Cj8O6w6jCizjDt1/DuRPDvMKTwprCvUbCssOLw6cJCMOfGD3Cr8KuV8K+XyPCgsK8SsOmMA0KGsKXwqcww5XDqsOEwqE8wqhMwot8wrrCqg8JCcKhGBPChcKyS8ObGyEMAMOCHDvDo8KwQWxRPj3CiMKbSShQw656HyzCpMKIwoIiacOOw7tIT8KZCsORw6AMfcO7w7jDpMKVOsKqTMOkHT90HVEvRiXDvcKOFUnDh8KEC8OLLWdOQcKRw4LDsT4nM8O2wqRwNS3DlUHDgsODK0XDn8KVG8Kxw5HDklQ3F37DjsKbehTCuF0UdFXDmcKIwrrCsVBuwrp1w5cbwpAmw4UdaXjCtQ3DiFfCiMOJQMK0w4NPwqTDrcOUSnZ2wrIVccO8woDClnoqDMKhe8OEUsKaD8OFw7XCoiJQA8OVM8K3wrhnUsKrwrl8B8OTLWt0w4TDhVlTAQXCqsOJw7DCmzzCicK4ASI5D3jCnXbClcKBw6kmw7J7wqhGwqwUw484wqrDijkoOxXCh8KMSlJNYFVaw5LDiSPCkkAkfsKfwqIcM0VOJE4UVRHDi8OIwqY5wqTCqHHDkRAQw6tRQsOBH8KaSsOB','bcO4w4Jqw7vDpMOqwpjDuXtte8OecEzCq8Kzw7MVb8Kbw54ywoHDm8KEFcKqwqkjYw==','w7t3bMKow5g=','HcObwqbCoDHDpsOpDsKhw5jCusKyw5PDtw==','wqHDjcO5HSM=','XMOOw7N/w4A=','w5/CqsOLJz0=','wq/DgMO7CT/DuxMqw6DCiMOGZMOhw4jCn8KPw4Fww6oKw7PCigZswr54wpB+wo/Co8OnK2xpwqhXw4HDosO8wqfCg1nCuEvCrUrDj01WeMKfYcOWF8O/cDVCwpBMwokAwpjCnkQIYcKWw6duw4c=','w5xwb8KXw6E=','HcKpF24n','wppdSHk7','wo/ClsO+fGE=','KGIlwpRv','w5NuGTjDsQ==','wqPCqMOYw4zDlA==','HcK7wpvChQU=','wqUiwpg=','JsKzZFnCsFbDpsKfw6NmQ2bDvMKFwofDjcKxRBvDjsKUJ1NLwqPDr2IHT8OHWMKcw4TDlxp8w4LDg8OCwpnClh/Dph4=','wp/DoiBISQ==','wr7CnMOvw7PDnQ==','VMO6wr4/w48=','w7bDt8OXFMK+wqwyB8Op','w6ZHwooow5Q=','wqFIQsOiTRQ=','FWsXf8OvaA==','w6wXEcOyNTgCOQ==','McODRA==','ezjDlcKWw4/DtsOfB8Og','QsOqw5BOw4k=','w6HChB0Jw5w=','w7ZyEcKRVQ==','EMKUZGY+','wq/Cj8O8w7w=','BsKOwpfCncOKTQ==','w6bCi8OADTdfw6nCuUfDhMODfhI=','dEvDjzPCgsKVw5nDt8OnBg==','w7RhKTkv','AMKxX2TCqw==','P8KswpQ=','acOjw5Ffw7rCrA==','bSsvL8KX','EsKMeEEp','wrhaw45ww44=','wrvCj8O6w67DkA==','woHDtk3CncKS','w7IXHsOyNTg4','w6DCrsOHGhE=','w5/CjCkSw4g=','w4DCnMOTA8OZ','KsK9woTCiysyw63DlsKLwo1KbcORfMK5wrM=','w6FLdcKew5nDgSDDlF1GHsKBw7YvIRg6F8KYMcOXGcKHwp8zMcKfwqBwBsKldMKhwq/Dk31oXm0sYsOuw50HSw==','I8KqwrDCnwg=','w7pLc8KHw4TCnGbCnUU=','wqPDr1PCnsKV','w7vCtsOTCQs=','wrXCvhPDjMKL','JXx7wp7Cjw==','wpFHwpMXCQ==','w75TS8KEw4A=','FmwLew==','SMOnT8KLw68=','w6xXDAE=','PsKmYlrCpg==','NcONBsKsAA==','ezjDuMKKcgFhwoHDng==','McK1eQ==','w5JaSGHorq3mspvlpKvotoHvvbPoroXmoKrmnZHnvbvotKfphLLorrM=','w6oTDcOlNQ==','w6RMZg==','wrzCiMORdUw=','ZCPDgA==','L8OpYQBT','dMOpQsKqw5Q=','w5PDu8KZA8OE','w68BGsOkGSQtMg==','bMOlw6JUw7A=','wppvbFgpw6M1wosSF8OHw5VawqHDkQzCmsOfSMOkw5XDksKRe3N8BBxvI8OwPMOHwpNww6VbHyDCgsKNw43CsEbCk8OVwqrCg3kcw6ZgXMOO','w7kCGw==','w57Cl8OfFghbw6nDpBfCqcKcOl8hVMOLEcOLOMKtw5PCj8OKwo81aglkaMKPwp9Rw6rCh8Ktw4rCoFRWw5piKMKGwqwswpvDrcKjw414wr/Ch3DCoWLChcKYNcKNwqECenQkISgAw75LSw4DwrjCmzzCtAnDv8OkwrY0HMOLN8KmwpcRwoBtGis/OsOCX8O4wrZTZQzCtEotwp7CkXkLwqEMI8Kyd1nDrD7CsMKXw5zDusKJa8KKwqHDsMKewrrDusOFWlDDjsK7w4nDjCkjcw==','dcO1woA1worCqE8tFz5ASydcw6/CilA=','TMOnw6JYw40=','w7cXUsO3ICNlN8OnNUrCt8K2','LnzDtMOCST40eydvBU/DtcOMAGvDjMO1XV/DqMKbWBE7a8KaOD7Cp8O7woUfwocwwr5FH8KGSDjCpMKFwo7DqsKfw5wZw7zDm8Ofe8OKw4rCi8Kbwo3Djw==','wr5nw55xw7s=','wpvDnR8=','w5Z/AyYMwpE=','wpotwr0Cw7k=','OcKowoDChC01w6k=','I8KQwrDDnMOW','TifDpsKbw4I=','w6gXC8O1Py4u','bC3Dk8Ke','w4LCg8O8G8O6w5QXw4A=','wpB6a00Twrd8w4s=','w5R5BwgQwoJsDg==','NMOowoLCq8Ot','wqfCgcOvw5jDh28=','fsKHw4kWwqk=','wpFKU1gK','Sh/DoMKlw4w=','5Lmz5Li/6L2f5Zuo5LiS56u/5pW55o2y','w7RMCwXDhDnCi0FPT8Onw5ZAHsKdQk3CrQnDrwtIEXRRSsK/H3bDmDxHwq8zwrZqGMKgw6zDrX3Cgyc=','w7jClwkvw6U=','LsKewqnCl8Os','LVcS','wqLCshfDmg==','w7IGC8OmI3BkcsOhfkjCtsO1w5DCvXvDl8KgwoVqaMK8WMKHw5cAGMKsw4xfw5plw4jDh8KnEQ9bOkrDscK7Lw==','w71/PSXDpA==','w7BdERLDg2s=','JMK6wr/CsjE=','wpLDh8OqCwLCoFFg','wovDvARzVA==','wrrDul7CrcKq','X8KRw5YMwq8=','bMO/w7p1w6/Ct8Kr','wqh3wqISMQhAXQ==','w44XOMOYFQ==','w6DCmkQMEsOxYeW8nuWnjOOBjeS7luS7lei1qeWMmg==','wq7DmsOrHDQ=','wpxye0MUwrh3w4E=','wr/DtnFrwq4Iw64tVU8=','X8KDw4U=','wrFCw6ZT','44G25o+T56WC44Cew4Vjw4EXWwXltqLlpL3mlIA=','5Li65Li66Lee5Y6s','eQbCtMKoIg==','PcKqwpDDvcOUFcOUNA==','XT/Dr8KRUgdlwoI=','w5Xor5TphYbmlobnm6/lv5Dojr7lj4o6wo5ePcKdw4ZiQ8O0RMKUwrdtV2lzHMK1w5TDj8KlV25vw7PClsKeS00Gw4DCmMOzTsOUwpfDinIewp0ZEsKsOQ==','woLDhBpWS8OhZDFlOBvCp2HCgMOoAwIiKSvCpcOpRA7Ci8O2XlNPFVnCmMOcLMO8woXCv8OZw4vDrcKpwpID','BsOSEcKlPsOY','LsK/cGghOHgeAxI=','fgnCvcKo','ccOgwoYuw4/DreW3meWlueaUuXIMHw==','woLDqE/CvMKMFXE3','5Lqp5LqQ6LaQ5Y+t','w7zCsj8kw7w=','w7EJZMOMw6c=','w7nCk8OAA8O/','L3vDjsOdXmE=','GMKPwrjCkQc=','KmfDpw==','NW3DrsOWdGtvPSl5','6Lys6KO05a+n5omc','wrrDqE0=','6L2F6KKQ5a+X5omR','5aSU5YO75pyC54Gf5Ye15LqX6KSG6L6m6LS35LuP','IsKodw==','UgLDqMKNw7E=','wprCrsOMeUI=','w60BSg==','w6o+5aew6LWcXknljrLlm5jDk1w=','dMOmwockw4rDpFI=','woFqS1w8','KcO8wpnCuz0=','w4QNK3d3wrLDoyp7XA==','eQwKSsKNwr49KMKDcMOsWg==','ZlPDni3CjsKpw5vDrMOqDAQhw7/DpMOoeMK5JXctwoJgw60/w7EpamN+YcOWwpAs','WsKEw5YfwrTDqzEKwpxUG8O5asKSw70DD8K7w7Y=','wqbDhMO/FSXCol1xw6XCncOMMcKmw5LCg8KXw4M5w60Zw6HCnUVzwqd9w5dzw4zDrsOiYC8=','wpbDnAo8wqfCosOCw5vDrjLCi8Kawo45IGfCt8OPw7DDsFctKVPCqMKiw57DnMK+wqfCjcO3w5N+GsKwAi5HecOaVMKtQMK4w4LDmTMjwqvDrQh2w6HDgsKTMEHDgT7CvTHCgW46f8K4wog4w4Ffw4J4w7gVwrvDhcKmwqoLwo9Pwr9iBMOTOsOVasODcnx7NsKjw57Cvj4OwoTDqiTDiMKrwocFQMK8w6sbD8KMw7zDq2AeexgRAVkGw6vDj8OULC8hOcOiR8OQw7/CshHDohvDt2xuw4LCtcOvwqg7w5MCIhfCnUotKcOewpDCv8OQw65rw4fDt2HDs8OAag9zHsKiZi/CjMO7eMKfYHNEw6NswqPCmsOJwpvCm8K1w6PCmG4RKsKiDcKAw7NMH8KawqnDi8KGwr3Ch8O1w67Ck8K4NxMBwqAcwr9/w7nDjDsgw47DsTHDucKUNHXDtMK+w63DumZMZnbCh8KWHhzCgsOlw5DCiyXDgmUOW8OBw7bDqWrDvQrCjsOXdmbDiid8M8Omw5nCs8ONJsKTw6zDiGFEw4/Dky9NVMK5Y3RjwoUpwrQGwpsiwrUlw7XDsUY=','b8OWw6zCs8K6woTCiEXCnQHCssKSVgNLw5PCsQ==','w7jCqxcHw40=','w792KATDmA==','NcKPwrHDsMOo','SMKBw5YfwqY=','wqPDv8O9Nig=','wrFIUMO5','wrzCj8Ohw6k=','N0UcwrE=','w69twpkfw6Q=','GMOANsK+','RMKyw648wrM=','wrDDlcOmDQ==','HMKpwrXCqcOK','wrvCshPDiw==','TsOCSMKcw4kxw5gLwpx5wrIMw6A=','AmoWaMOcZMOHU0YBI2cpwqhCw6rCgzQYY8Khw616HxMmd8KnbcKLCG/DuRg=','c8Oswp0sw5DDoV8xODYc','w7zClBkrw7U=','JMOPwrvCvcOswozDqFDDhg==','T8OFdcKCw7M=','w4QYMzNg','5aW45Y2H5Yif5YqB56KJ44OOw6U=','KUPDucOocA==','wrF/wqgN','w7fCl8OvEAY=','O8ObHMOq','IsO9KMKjU1tFYsOr','OsOSOsK4FMOcw4kN','GMKBwpLCk8Or','ZwnCucK5','wqs0wqInw6XCug==','B8OdwqLCvws=','w6VDwpY/w79ow4rClFrDrQ==','w7JZEhA=','5Lqk5LmQ6LaB5YyI','w4lLGgfDuWLDiQs=','wr3ku4TkuYPorq7igIzmn7Hmtrfli6nkuJzkvZnml6vnvpnvvarorLnlha3ms53lhL/kupvmtoDliqnjgI/igoc=','wo3DocOkFyc=','w5AnFMO4Ow==','w7IaX8O9w4XDp8KdZX0=','I8K0dw==','5Liq5LiP6K6X4oO65pyc5reC5Yqo5Lmm5LyV5pa257+1776B6K2G5YaE5rG95YW15Lmt5rW75YmM44GW4oKU','w6fDpsOEEw==','FnIMTcO0','wqrDpsOOOz4=','C2Yawqx/','GuS7hOS7jOi3q+WOrQ==','wp/DvHthwqTilLbojbPlvpA=','ainDhsKR','eQbCtsKsCA==','BcO7dAtb','DsOAJMOTw7M=','w4AUTMOlw6w=','KmLDs8OhQg==','wonDiw4+woTDscOKw7nDox7CisKbw5A=','w6JyF8KD','w6bCtDozw6F9wqdoGyA=','Ti3DpMKoVw==','wolJUsOORQ==','KcO5wo0=','w5B1FxbDmQ==','wp/DpcO2DBQ=','woJ6als/','w7hZCxQ=','w6nDrMOC','55eu5oip44OR','w7vCtTgqw6pDwqli','wpjDmR8t','LcOFTQdsFA==','YcOtw4J7','JcKkwpfCjAwww6HDhw==','M8OLwqzCrg==','Cmwf','PMKcwoDCsCA=','AsKfwpvCnsO7','WsKEw4ALwoI=','PEzDocOGWw==','wonCrsOKaU8=','wohJfVs=','w4pxFhAb','fAfCtw==','FMOcbhl0','LsO6wrXCmsOz','w4jDh8KaB8OC','w7lOdcKsw7g=','wr/Cvh3Dj8Kr','AUB/wpbCucKxVlJwOcKWXcO5w77CnlPDkh/CrAJpwqLDmTbDtMKuw5lYwqLDinLCgcKuwoc=','w5HCoMOSCMO+','W8K0w5QGwoE=','BV/DhcO5Xg==','wqkiwok4wqzCvl7Doy3DrQ==','w5nChCIow70=','w4zDhw9fBcOrbXJpOkfDryPCjMKyVEB/IyDDtcOgUwXCtcO5A0VHTw==','w4B4SQAQ','wrVYX8O0VQnCmCDCji5ywoscHsOWFx/Dk0/Cp8K9MMKCAUs+w6A4LsKzw4HDmMK2woczHn3DisO0w4lRLBTCjzls','wp3CnsOREMOAw6vDtHoCegITw7zDvDfDvAzCi8O8LMKUw4XDmyxKw7vDmxZ+QMKNDSTCpcOTw4Rpw7oVBA==','w6vClsO6CinCs2xsw6/DkMKYPMOuw5zDisKawoNww7wSw63DlB1rw746w51xwonCq8KmO1Niwr1ew4DDpMK9w7nDmhbDpgLDuA==','wonDtz5IXA==','RcOVwpgIw7E=','OcO1ZDZI','NcKofVom','H8OOLMK+','wqMFwqYDw5k=','w7/CjR4Tw5c=','ajrCtcK+','w7MBMcO5NC8=','fsKWw5IBwpU=','w5Qvf8Osw7g=','wpTDkcO7VA/CrlNuw6XClw==','awwPDcKU','w4lgCAoK','M0gAwpJD','Wiw2McKa','JMOewqrCpsOnwp/DkV/DkQ==','OcK7am0=','wqDCthTDmMKMZQ==','wrzCog7DvcKq','KMOXM8Oxw40=','w5DCkcO0DMOgw5kew51p','DsKKwo3Cmw==','OsONTitSEA3Chwo=','SjbDtsKyw5M=','w65ebMKLw7nCmGDCiVk=','6I215b+Qw53mrannvbrnipPmnIjkvpA=','HsK5HGM/','VsKRw5YO','wobDlQBBTMKz','KEYMfMOe','wrHCiMOMXmw=','wrZvV8KeAcKFMxRH','wqVnw6pCw7U=','w55xEAI=','DsOCK8KDHg==','L8OpwqjCjsOi','O1F4','MkAPwqdB','w6FHwpEv','wrbClxvDi8KZ','bMOEw7Rww7k=','bCXDicKYw5vDucOfTsO6BlJLw73Dhg8yw47CviwfwpDCmsOlw44PSS/DhDE=','ecOrYsKqw50=','wrxCw7lSw5jDrHpo','wonDkRxCdMKyOGo=','BVxRwpXCiw==','J2bDs8OFX3Y=','KMOIFMOp','SMOzSsK2w6A=','PMK5amUZPngOLA/DnA==','w6HDlcOsDSPCs2lww6XClsKf','w5RrcUZn','R8K3w7A1wpY=','QcKVw4EdwqLCpU5Mwpo=','csK/G3MxwrLDtFI=','wqLDiMKVTg==','wrc0wok6w4jCsVTDpQ==','Okc+wp1P','IMK5wp3CgsOs','w4QIYsOYw6Q=','OcKzcGsVP2VYBgPCgHh4w5PDhMO5w6/DosKic8OpwpfCnCNGL8KDw5HDt8Oz','wpN4bEEswrBuw502FsOX','wrEvwo06w6TCnF3Drj4=','NMO/woArwps=','M0EWwrdrw58Dw5nDrA==','QA7Dv8KGWA==','wrTDuljCqsKOHW8m','ZCnDpMKEaA4=','woHDnipFWg==','woJ0a0EuwrB1w4o=','wobCncOYw5TDnA==','wpFxwpIaNg==','QsKFw5EH','fMOmwoouw6jDqUYt','wrNtwqQLNgdLVw==','XMKZw4EEwqnCsHNA','YSLDg8Kaw5nDnsOQ','PsKoY0DCtwXCpsOe','wrVuwq0QHAw=','w7pDwpY8w4Vv','X8O8wrkMw48=','J8KiwpM=','5ba057uH56+U5a2L5om65pyN6aKD55q65LuJ44CC','BsKEwp7Cv8OMVw==','woHDgU7Cr8Kh','by3DisKaw7LDssOZE8O8','w5PDl8KdAsOHw6Y=','w6B6D8KWVcKL','w7QbRMOw','JMOkwoDCqiY=','H0/DgcObfg==','w7R2F8KBX8KdZQ==','GUddwoPCvQ==','NMOfbyFmGgw=','w7dWOxbDlQ==','w4UNKzl7wrnDqQ==','bcOtw4VVw7/CsMKVw4XDunFsc8KHYA==','woVGwosRLg==','H8KYwpzCiMO3SxzDlA==','CMKKworCn8O3SxzDlA==','w5kBPDF6wrzDoSQ=','IMOVEg==','5Z215p+25Zyu6aKU5bu85Lm65omy5YmH5Luj56y/5qKp77yM','wqbDl8O7EDrCqEh8w4XClsKf','NgnCs8K5NcKQwoo3wo/DmcO1','axQCFsKFw5M4KMOI','TMKbwpDClMKD','CnMDwqZi','w7DDtsOMGQ==','X8OiesKaw54=','w4zDgAFVUcKvInFpYA==','w7LDosOMCQ==','wrzDrEjCqcKs','RcOOwo0dw7M=','QMOJwoUVw7U=','w51ZTcKdw5s=','EcKuwpTDtMOS','w6bDscOIUsKgwqo8BMOGLidDwoRDw67Dg8KZw6AeGWsow4ZeHHdaKQticsOd','w6wpWMOFw4A=','JE0bwqJ0w4M6wp/DocOlw4nDlMK+wqbDlyYtUMOpA2rCt8ODKsO1VMKxUg==','wqrCicOmw5XDhw==','fBUNA8Kaw7g+Y8OOe8OiWQMuU8KJGGfDhyhAUMKzbcKvwpk5wq4nGcOXc0YHdw==','c8O2wpk1w78=','IsKWIHk6','woLCmw7DvcKP','S8KRw7MBwrA=','G8OKAsO5w5A=','JcKqwrXCsQM=','G8OODMKiNcON','woDDmV3CnsKp','wq5Zw5piw5o=','wrdCYcO+ag==','wpPDkT9ITw==','w5jCqxQ3w60=','XzzDvcKETQ==','ZkTDgAnClQ==','wonCoMOLbg==','wr9/wpAXCA==','IUcBwqx4w4Inw4nDi8OpwpU=','HcOBRMKAwqc=','w6AYKD1F','aMK3eUfCigHCrsKN','w7XDqsOLNMK9wqw=','w6vDqsOGFsKewqo2BA==','w6rCsBDDhsKAXXgADGzChcKsb3I1w4fDont3w4DCjkZ/w6A+wp0AK8O1','w5PDncKU','P8OOB8O3w6/CqTxzwoc=','I8KiwoHDpcO/','dUrDnSo=','dBME','5Lib5Lm06K6r4oGC5py45rem5Yik5LmP5L2y5peZ57yT776N6K6R5Ya95rO25YWD5Lqf5rW25YiT44CD4oCU','WMOQX8KKw4k6w5YLwpw=','w7HCvS8g','JMOCwrnCvcOswrvDl13DjQ==','NcO0PcKw','JMOywo7Cuy7Dpw==','CFBvwrXCocO+','f8OuwoArw6fDq18hBzc=','AmIMbg==','w6jDosOME8KRwqgvCMOmIg==','wp7Dnz1OV8Kr','wod+RsOHaw==','w5PCkcOtCA==','w6FyDsKHY8KabxoQ','w5HDo8K9CMOy','wr18f8O6YA==','IcO5wqDChyc=','bC3DvsKC','cQ3Dv8KMUQ==','woXDuR4jwpo=','LEsS','wq/Ch8Omw7rDj3UCZsO9wrp6w6ouEcKywpvCg0gqBcOtwozDmm/DpnhQaEVWIcOB','woHDvsOmKwg=','w7YdGA==','wr3kuZzkuoXotY/ljYQ=','KUoRwqB2','CMOfRjxPEg/CkA==','wpwGw5h7wpHilJPojInlvao=','wrtRw6pBw53Dq29z','M8K7c2k=','w4dgwpM1w4s=','w51zwoEUw5c=','BMOJPsKFKA==','wpTCrsORdGsoUcKRw5os','w7VCBsKqZw==','wr7DlQwuwp8=','wpHCkcO6HcOcw4gkw5plVcKx','LjzDo8KNIQ==','w6rCpxvDjMKTWWgeVHc=','wrjCqgrDmg==','ZMOuwoUww4M=','J8OwJ8K1U2pjYsKz','KMObAcO/','worCp8OXamMv','w4M6a8Oaw48=','w7PDosOJCMK1','d8OlIMK/Cw==','wpnDlQ1UXcKvG3dp','woZiaE0=','EsKLwoHDscOz','wrvDtE0=','5raY6KWT5Lyw5Z+0w6JZPA==','w5kJMj8=','w75eaMKa','HcOMJcKLEg==','OMK5wpXCkzci','wo1lw65iw50=','w6bCv8OTM8Oi','wqpxwqY8DRs=','wpdcwo0yOQ==','wqjCuhTDmMKCZXhBUiLCh8KoYG5Kw4/Drio0wpzClUZ7w6A/wokaJA==','w6VQZg==','TMOQRMKa','M8KYwpLCqTA=','w45IdcKWw4U=','T1VowpLCpcO5LAh4LcOC','YMOWwqQww7w=','eynDqcKReRJYwo7DiQ==','WsOVScK9w7Es','worCqsOMbkMlQsKL','wqd6wqUqFBw=','wqlCw6dDw7E=','IMOVEsObw7PCvA==','J8KoCGg7wpzDuQnCgA==','d0LDnDLCgg==','EMOmwrrCuMO9','w6fCl8OOGgo=','J8KswpjDs8O0','woDDtCJWYg==','BsKdZm3CiQ==','w53Ctzo1w6E=','YCnDq8KHeRR7','w61+EzAc','OcKCwr7Clwo=','IMOTBsOq','GMO0aQtN','ZMK5w4UmwqE=','w7EPa8ONw58=','5Lmu5Lmb6Ly45ZqW5LiH5Luf5q6H56uw5pSr5o+m','Y8KZw7Qfwqo=','YQnDqMKtw4Q=','MnQUwqln','5Lip5Lqf6L2x5Zqj5LmW5LuL5q+u56ib5peg5oyp','w7RfISs7','WcONwpgKw4w=','H8OQJcKLEg==','EsO9K8Koew==','w70HNMOQGg==','wrlvZsOibA==','w59vVsKPw78=','E8OGwr3CiRA=','QsKRw5AcwqI=','LmF8wrPCkA==','YcOqwp1ow4XDp0QjGDc=','TxnDkMKqw7s=','HsOiAcOIw6U=','woDDnwdI','ZmTDmgnCsA==','P8OiwoTCqwI=','M8K9F2QGwpjDvx3CnA==','w49RCzHDhQ==','wqtpwpERGQ==','w5TDk8KLJ8Oq','cwnCosKpFsKLwqw2','PcO8OsKl','wrRMXMOycgPCmDzCog==','w5vDk8KHBA==','CMOAMsKvCcOew4safg==','E8O8McKpZw==','w5jDk8KeAMOgw63DqUYG','w6p8BA==','6I+j5b2LS+asoOe+rueKt+aequS9tQ==','wrtMQsOYVg7CpzzCqDoqwpodEw==','esOfe8Kvw5c=','PsKGHnQf','wqTCjMOEw5bDrA==','ClV5woLChsOiCgk=','C8OAK8Kr','w7pPwosv','wpHCrsOLVV0ldcKKw4M5w6/Cu8OFfQ==','wqvDm8Oo','5ZuO56y85q6z56Kj44GY','MMK9DmA=','w5JxFywJwo1RGUfCqMOKY8Ouwqs=','BMK9NVcD','YSDDqMKHw7U=','TmzDqTDCtw==','JMOxwonCixE=','dBrCscK6FcKJ','DUZqwpHCg8OlHxI=','OcKof3smOWoY','dsOuwp0k','AnEZeMOvYsOIE2YZIGk=','wqPDlcO7GA==','bD7Dq8KUVQhuwog=','EMOlSg9q','w7kvQMO8w5s=','YOS5h+S7pei3nOWNiQ==','w5bDnMKXAMOL','wr1EUsO8bwHCmis=','w4hnw4xowqHil4vojoXlvJ0=','w5HDk8KeAA==','a0zDiQ==','5oqv5aeZ57i65p+Zwpw=','XsODX8KBw6gUw5wKwopbwroN','wqBdXcO+VQ==','wp50fw==','wrpRw7lZw6bDiGxvQ8KHTSw=','w4TCiwkEw5c=','wrvCj8OOw4TDgQ==','wobDnwk=','w67Dv8KwBMO0','dEvDgTHCrsKu','TcOUQ8KKw78rw7Ad','w5vCn8O+','V1rDtxLChg==','wprCoMOcfw==','OsKoe0zCrSfCrMOJ','JcO6IsK0WFNPfw==','wqMgwo8rw5U=','wq4owos=','KcK1dWkB','wrZMbXEM','w5JkEBMNw5kuREnCqMOGP8Oow6DCnSwmw4dAwrLCp8Kzw5XDg0/DpMOew6wnwoIzwq/CjWhmw4zDgQ==','w5YYLzZ9wr7DrTV7VmXCrgzCvXFb','wqTDqRRTUw==','w7TDssOwEMKl','wrIowp88','PMOgIMO8w7Q=','w5hIDg/DtQ==','YUXDow/Cvg==','w7BXGA==','P2NZwqLCsA==','wpUtwpo+w7c=','w7/Cl8OC','55ap5omZ44C9','wqJ/wrUY','woTDmQ1NVsK6Jns=','44G35L+i5oCX6I645Y2w5oic5Yix','fT/DgsKNw6jDtQ==','wq9Kw6V/w7nDog==','w69TwpYWw5hjw7fCkF3DsyvDrlvDjA==','w6h6AMKJfsKYbQ0=','CGobZMOHYcOL','wo/DrsOJEC4=','w7dYFDIU','RsKfw4kKwqk=','O8OFwr/CisO7woo=','NXzDssObVGNyMjY=','eCPDucKX','Z8O4w7lqw4M=','bx8IEMK3','w45TTsKqw4Y=','PsKoY10=','wpDDmsOiExw=','w73ChsOXKMO/','w7xMZMKcw6PCnzI=','w4TCmMO2GcO6w54=','LjjDpcKIeQg1','wqtMw6BTw7o=','HcOXX8KBw7cNw4AJwpwHwpw4w5U=','Hk58wqPCpw==','T8OLQsKYw6M=','OsKQZ0HCpQ==','BsKEwp4=','wqBZQ8O+TwfCnijCvg==','woVhb203','M8KzwqPCqsON','AcK7Imgf','FGYLesOKeA==','SRnDnsKrQyVXwrLDtDnCkD8=','wqNMQ8OkRA==','w5ZyEhDDjQ==','OcONVy8=','fT/Dr8KRTw5pwpXDgj/CrWbClg==','J8K0G3MwwqTDswDCncOy','IcKvH3MGwpPDsR3CnMOUe8KnNA==','BsObJMKJQw==','w4LCv8OqBSA=','w5vDiMOcBsOcw6PDq1sNOABUwrLCjXzCvV7CgcO2AcKbw57DpRFEw7vDvVs=','GVV5wpXCrw==','MMK9U2YW','w6HCszAkw6ppwqF+','w7HDrMOOGMK+woA+GA==','J8OFwqvCuw==','TDUhHMKP','wpsvwpoaw6Y=','BsOQMcKiQg==','w69hFwYa','w6IZLD9w','w5Z/Aw==','wrjCmsO6w7TDm3oCL8On','wrR7wrIMEx0=','w7bDpsOGD8K1wr8LCMO+','w61edcKP','IMKmwpDDpMO/AMOpOMOI','LsOJQDxkBzLCnAE=','w4Y8bMOFw5I=','wr7CncO5S1M=','wqxXw7lfw7rDomB6SQ==','J8OLwqrCvMOs','w65dDADDm3c=','wqBFXsOnaAQ=','fB0XBQ==','Zg3CvsKpP8KQwpYm','G8KSVUvCgQ==','wqjCj8O6w7nDuXQYPQ==','w4plFws=','wqLDvD5waQ==','A8KFwp3Cn8OGahw=','w7lQcsKHw57CkmDClQ==','fQ/CncKnIw==','HMOtL8K5HA==','wpnDvB5Vfg==','SQDCpsKfPQ==','wplUXsOfbw==','HMKWHEk/','wojCtsO9ckw=','w7kGNsOmCQ==','WcOswoAKw6s=','KsK9wp3DiS9/w6bDhsOMwoFLLw==','w7sCD8O6OSkqKcOqdEfDt8KjwpDDpGbDhMKjwoBqd8O+F8KXw4QCUsKxw4ZXw5BJw4I=','wr/Ds3E=','wq17wqQJUghBUcKfGQ==','wpDDmENuWcK1ODNEE0HCuHLDnA==','WsKEw5YfwrTDqzEKwpURXMK6KcOWw7NJQsK3w7Q9wrdYLB/ClULDiMKnR8KKH8OzGmN5w7xkw79lw7d2woXCnjJZwrtTRELDs8KkNMO9eQ==','UcO2w7tRw4M=','w7ElQcOxw7E=','HsOpwprCu8Og','w6YURMOkwofCoMKQZmLDgg8Xwo8VwofDucKm','wqAowogxwrzDugXDiH7CujTCj8O9Z8KewqpBZyBjw6pBWsOMw7ANYl4zw5gzwpg0J8Klw5/CoQLDp8KWwrjChVpoLMOtb8O9WULCicODMmc8w7Z9w63CjcOVXmo9w70PwpTDmMO+I0nDjztqwoTCpRwABBcmwqrDpiPDncK7w5Jrw5omw5ZCw6PCg8KDOsOcwp7CqMKrNcKuw71NWcO9w7Njf8OyMnLCmn81Zw3CqW3DgDjCisOmw4xDdcKfwpHDgi/DgcKqVMKdwpbCi0d8WgHDisOBJcOwWcK8T8KiwpUiTClMwrvDjsOyN0B6w5N5wrZ6w6PCri9TOMKmwqcuYsO9wq7CvDvCgQLChAkTKzPDpB/DucKfNcKNw5MqGQjDjy5dw6NmwpLCl1zDnMO9woQ+wqNewrzDssKnSTc5Q8OCDcOSwqPDgsOdw6UCwqbCicOOaV3DpMObUwJDw7nCmcKrw4vCscOTY0DDhcO6w5PDtwzCsEbDrsKeM8KKKHMrYnPCsxpQHi0vSQsRwqgiQ8O1wqATeVLCnxg8wohDMcK/CMOGwqnDn8Ojw61KTcKgZH8eYHLDkCREU2DCoA/DhcOow7rCliYhwobDsDMEw6UnaMKywqTDpkPCusK3wo8kw4kwwqHCj8OUSAZoT8OXwr96wo8FPzB3wq4GWcO0VMKpAjRfCEBueB7DsMKiHcKxw6wvw5/CrnJew40Ww7diwpLDnxAzFhBRKml+wrZjPsOncRspXMOoZcOzw6HDrQbCnzLDlsOUJgg8wrLDnsKYwqpJw55jw5nDlAdEw4zDusK6wr5YdcO4Q8OUKsOcwrc7w5DDoMO7wp3DvcKxRcOwwp3CowPDvW1JcsKiUiIEFkQbEXEjRVjCon7CiSvCqsOgw6jCgcOUQCYCw6AOfsKbwpkuUCVtw6kXwoPDq8OMwpXDnCHCncKJw4zChsK6UsKLw73DtcK2wp4/w4nDujhLTUwQw4h2FkrDlCbCr8ONw6fCpMK6w6cMw7bCvFbCtsOHAMKOw7fDj8KDw5wsw6QXw5p/w5c9QMOWw78rKQ7DqRzCg8Kgem/DrEHCuRpkEXnCtATDlcKEXGjDs8Oiwp/DuzbCrV8rwotUwq/Ci15gw77Dl8Odw5XDkC/CvcKVE2/CnAXCiAAXM01mTSE8CwNfw71KBigDdsO5ZHQ2w4bDgMO7w4nDqMK/cAodwqrDscK6GcOOesKgw7jDoA1OQGE+EsOAwppGJEwJO8OZegDCpwHDlxHCjQgmYMKVwqQlwrPDjAnCog3Cp8OGLMKlesKAw7A4wocIX8OZLiR+HxzDlcKQw5HDolAxJMKAwoPCksKfcsOhZMKYf8K3w63CoXJLw4RFdGjDscO4McOyw65hw7BJR8OnCiF/RMKrw4o6b2FJwo1aw4swwqxYfQsJw7ddcsOQw5EKw54Bw6DCnDjCoSvDkGDCrCjDnsOEMsK5BsOzw4jChRYZw4/DncOqw5PDgGfCiHElbcKHwqzDmcOORsK3w7UEOAfCsBfCg8KpwptrYSMXc8Ojw4nDn8Ovwp/CmGAHwrXCtQrClMK+w4APdwInWDHDoSTCucOOJcKNAcKNLHZdwpfDiBEEw5Bewq/DiMOCwqjDoBzDiMOHw7Fcw7AXwpEIwp5ETwY1GlUsNcO0w7o+wrXDv8KpWcO4wrbChsKALcKjABPDqMKBwqPDqTvCkBFvPMKxRcO+wqUAw5oYwofDiHvDtnsvHTB3QsK5wqHDlE/CvsKxw4Jbw6gCw7JXw7BEw7nDhcOIfcKrNMOSQjzDssK6wqY2W0LCuDrCsGd/cF3Dv2dhKQ==','w6ZJwosv','VkvDlw/Cvg==','w5MBCcOCPyEuM8K+','ZAfCu8KoNMKpwro7','GcKfwovCk8OQQhPDnQs=','O8KswobClCc=','BUvDj8OAeA==','wqnDjCI0wqY=','w64dFMOzPgEuJA==','wp7DqFzCmsKtH3k8Kg==','w6fCl8OOGgp8w63Csg==','cgjDq8KXfQ==','S8OQX8Kdw78=','w67DmsKKK8Oq','AsO4wrvCrcO/','w7fDkMKnIMO2','w4TClcO6G8OWw44hw4Zi','w4A7ecOcw7TDg8KrVlfDqzxe','w7V2AMKQVcKNUAEb','OMKzHQ==','wqzDjC0WwpM=','DcKuWFYr','WMOeX8K0w6s=','wrPDscOHDCM=','IsKlFFYS','F2AewrN3','b1fDmjHClMOwwpXCt8OiEwMgw6rDp8O1a8Oga34vw59uwqwjw6YreyN8bcOGwpwnWcK4wqXDtsKWwr7ClMOmRMO5RWwKwp11wpd9FsKeLMKDwrnCshhfwpk=','NcKsCm08wpjDsRvCkMO4esOsKcOUHcKgQj1ZwqJUw7XCoVhPe8OxZsOUEiBnwrw=','wpfDnQ48w7rDuMOHw6LDsDg=','bzbDjsKPwo3CscOSBMO/AlJRw7/CjwAxw54=','w6XCpCsgw54=','WsOBRMOAw7d3w5Mdw5dZwrIF','MlNJ','wplpBcO+cQjCmCDComV+w55eX8K1XVbCmEPCtcKsCsKfAQI4w6sYGcOqw43Cl8K6w5VrSkfDn8O8w5FNSkPCgzN+Qw==','w6RcLMKIaQ==','wo/CvMO+dFk=','w75XGwzCiibCkywIGMK0w40cH8KWGhvCplnDgUNVQXlEUMOgHzrCjBMrw7QUw7YgcMOkwrrDmjHDng/DusK/w4Rhw43CssOIYcK3w57DqnXDqTwZTcOISGBhP3Ayw4/Dnws+J8KAMcKBw5www6fDn8OkwoDCuVMJe8KWwqRmwp1bL8KAw7vDqMORwpZUOMOJwrvDrQDCi8ODw6hyE8OGwqTDuDFQw7bDgsOkw6QlbztCw5XDmyY9w5DDtsOcwrhsw7/DpcKeMX3Dr8Oww6DDjgvCpsO1NsOSw5vCn2FULGs7w7hXw6/CmnbCssKJwrTDuFgqMMOTdsOCwpZZX8KlO8ObCcKxdkQeW1fDq3fCtcKZw6PDscKWwqLCuGPCuSY4w5godiIFNcKNJcOzw7HDnRJNw4nDhcKoFg/ClsOOwrDDnGjDvxnComoQwqLDiF95V8KYw4kYwo3Dvw5dw4M2wrjClEtRZ8ODw57DnMOIwoDChFF9w63DngUP','wrZxwrIN','w7BxLyHDsw==','w4VsQ8Klw6I=','MsOMwobCvzU=','w7pJwp8=','M1AHwqxgw4w6w5bDuw==','wr5IWmMS','HXFDwpPCpQ==','BMKOdGrClg==','L8KKR2PCgg==','wqEowogt','wp9+c8OcaQ==','BcKme2PClQ==','wqovwrsxw40=','woZ0c000','wqAiwo0m','wqjCsg7Dng==','InrDocOFc2p9Ow==','J8OzwovChgvDp8O0','eQPDhsKlbw==','wrfDmR4owqA=','IMK3woHDv8O0E8OQN8Of','w5ZZNDc6','UcKEw7UZwqE=','wo/DjxE6wpw=','PsOOMsOuw7c=','wq57wqAdGhte','eQzCncKjIg==','DncMf8OVNsKBU14aNHZlwqJFwrPDjjIGeMKkw6A5FBQjLsKHbcKIU27DvgLCmMKhScKXCDhGQsKFwrAHWSMGBcK0w4IFOWMuwqoJT3DDoH1BFTXDq8OIGcObeFJf','wpcqwqgww5U=','OcKnwrDDm8Oy','wrdaw7t5w7I=','A8OiwrnCnCQ=','w5ILWcK5w6jDr8KbaG3Diw==','w5nCuMO/DMOZ','L8KowrPCvzI=','wqDDkcO7','w6ZiwpQaw5I=','w6DCjMOXFgpQw6HCrVs=','wobDnwljSsKp','w6hwCcKYeg==','w6lCCi8G','GsKtwp3DsMO5','NMKiwp7Ds8OJF8OWI8OD','w6BMT8KBw47Cng==','w5/ClcO4DcOWw4gC','AcOCNcKwEA==','w5R5EQHDpw==','B8KpwqvCuMO5','wrvCnMOxU3g=','wrE3woAhw7U=','wrTDhMOjEDg=','wr7CkDHDnsK9','w69IExzDgw==','CGASdcOs','HcKLwrnCrgw=','DMKWQ1HCpg==','SMOFX8KHw7Q+w5AfwoA=','IsOlJcK4Qg==','jsjiaWmJKkZi.ecDotmM.vxd6PwQquQ=='];(function(_0x3913b1,_0x2a6020,_0x157b52){var _0x56dfbc=function(_0x23395b,_0xc8c034,_0x11735f,_0x211dfd,_0x37d3db){_0xc8c034=_0xc8c034>>0x8,_0x37d3db='po';var _0x40708b='shift',_0x1dd4e4='push';if(_0xc8c034<_0x23395b){while(--_0x23395b){_0x211dfd=_0x3913b1[_0x40708b]();if(_0xc8c034===_0x23395b){_0xc8c034=_0x211dfd;_0x11735f=_0x3913b1[_0x37d3db+'p']();}else if(_0xc8c034&&_0x11735f['replace'](/[WJKkZeDtMxdPwQquQ=]/g,'')===_0xc8c034){_0x3913b1[_0x1dd4e4](_0x211dfd);}}_0x3913b1[_0x1dd4e4](_0x3913b1[_0x40708b]());}return 0x79963;};return _0x56dfbc(++_0x2a6020,_0x157b52)>>_0x2a6020^_0x157b52;}(_0x516b,0x6e,0x6e00));var _0xf668=function(_0xcd51c9,_0x3aeefa){_0xcd51c9=~~'0x'['concat'](_0xcd51c9);var _0x3256e3=_0x516b[_0xcd51c9];if(_0xf668['rkprnE']===undefined){(function(){var _0x243ed4=typeof window!=='undefined'?window:typeof process==='object'&&typeof require==='function'&&typeof global==='object'?global:this;var _0x3a199c='ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=';_0x243ed4['atob']||(_0x243ed4['atob']=function(_0x38afd0){var _0xa98c71=String(_0x38afd0)['replace'](/=+$/,'');for(var _0xdceb86=0x0,_0x372f20,_0x35ea52,_0x3fc5be=0x0,_0x4344e9='';_0x35ea52=_0xa98c71['charAt'](_0x3fc5be++);~_0x35ea52&&(_0x372f20=_0xdceb86%0x4?_0x372f20*0x40+_0x35ea52:_0x35ea52,_0xdceb86++%0x4)?_0x4344e9+=String['fromCharCode'](0xff&_0x372f20>>(-0x2*_0xdceb86&0x6)):0x0){_0x35ea52=_0x3a199c['indexOf'](_0x35ea52);}return _0x4344e9;});}());var _0x386ad6=function(_0xd08aae,_0x3aeefa){var _0x232725=[],_0x1dc0fc=0x0,_0x1a93ce,_0x485626='',_0x8f1d48='';_0xd08aae=atob(_0xd08aae);for(var _0xf78750=0x0,_0x4d36c9=_0xd08aae['length'];_0xf78750<_0x4d36c9;_0xf78750++){_0x8f1d48+='%'+('00'+_0xd08aae['charCodeAt'](_0xf78750)['toString'](0x10))['slice'](-0x2);}_0xd08aae=decodeURIComponent(_0x8f1d48);for(var _0x340ccf=0x0;_0x340ccf<0x100;_0x340ccf++){_0x232725[_0x340ccf]=_0x340ccf;}for(_0x340ccf=0x0;_0x340ccf<0x100;_0x340ccf++){_0x1dc0fc=(_0x1dc0fc+_0x232725[_0x340ccf]+_0x3aeefa['charCodeAt'](_0x340ccf%_0x3aeefa['length']))%0x100;_0x1a93ce=_0x232725[_0x340ccf];_0x232725[_0x340ccf]=_0x232725[_0x1dc0fc];_0x232725[_0x1dc0fc]=_0x1a93ce;}_0x340ccf=0x0;_0x1dc0fc=0x0;for(var _0x3f62f1=0x0;_0x3f62f1<_0xd08aae['length'];_0x3f62f1++){_0x340ccf=(_0x340ccf+0x1)%0x100;_0x1dc0fc=(_0x1dc0fc+_0x232725[_0x340ccf])%0x100;_0x1a93ce=_0x232725[_0x340ccf];_0x232725[_0x340ccf]=_0x232725[_0x1dc0fc];_0x232725[_0x1dc0fc]=_0x1a93ce;_0x485626+=String['fromCharCode'](_0xd08aae['charCodeAt'](_0x3f62f1)^_0x232725[(_0x232725[_0x340ccf]+_0x232725[_0x1dc0fc])%0x100]);}return _0x485626;};_0xf668['NeehNH']=_0x386ad6;_0xf668['RaKUFm']={};_0xf668['rkprnE']=!![];}var _0x5182be=_0xf668['RaKUFm'][_0xcd51c9];if(_0x5182be===undefined){if(_0xf668['CLbpqE']===undefined){_0xf668['CLbpqE']=!![];}_0x3256e3=_0xf668['NeehNH'](_0x3256e3,_0x3aeefa);_0xf668['RaKUFm'][_0xcd51c9]=_0x3256e3;}else{_0x3256e3=_0x5182be;}return _0x3256e3;};!(async()=>{var _0x326225={'ZNOrP':_0xf668('0','kmgF'),'aGBPS':_0xf668('1','%W95'),'Djkau':function(_0x5626ca,_0x45ad70){return _0x5626ca<_0x45ad70;},'owKUs':function(_0x2ffb06){return _0x2ffb06();},'aLjUl':function(_0x432867,_0x49205f){return _0x432867(_0x49205f);},'TeGNE':function(_0x3dfaad){return _0x3dfaad();},'nXFgh':function(_0x84576b){return _0x84576b();},'pgIXL':function(_0x1977c8,_0x4d3e9e){return _0x1977c8!==_0x4d3e9e;},'NcYjL':_0xf668('2','vPrL'),'SBLvE':function(_0x52ca1a,_0x2e6d62){return _0x52ca1a===_0x2e6d62;},'VRpbB':'vXAbn','KRBnA':_0xf668('3',']AF0')};if(!cookiesArr[0x0]){$[_0xf668('4','uhqA')]($[_0xf668('5','yzvA')],'【提示】请先获取京东账号一cookie\x0a直接使用NobyDa的京东签到获取',_0xf668('6','yr@1'),{'open-url':_0x326225[_0xf668('7','%W95')]});return;}for(let _0x1a771c=0x0;_0x326225['Djkau'](_0x1a771c,cookiesArr[_0xf668('8','%W95')]);_0x1a771c++){if(cookiesArr[_0x1a771c]){await _0x326225[_0xf668('9','W6#w')](getShareCode);cookie=cookiesArr[_0x1a771c];originCookie=cookiesArr[_0x1a771c];$[_0xf668('a','C@fO')]=_0x326225[_0xf668('b','s^UP')](decodeURIComponent,cookie[_0xf668('c',']l!6')](/pt_pin=(.+?);/)&&cookie[_0xf668('d','V[6Q')](/pt_pin=(.+?);/)[0x1]);$['index']=_0x1a771c+0x1;$[_0xf668('e','nKU7')]=!![];$[_0xf668('f','#IEW')]='';await _0x326225[_0xf668('10','yr@1')](checkCookie);console['log'](_0xf668('11','s^UP')+$[_0xf668('12','C@fO')]+'】'+($[_0xf668('13','[XoU')]||$['UserName'])+_0xf668('14','vPrL'));if(!$['isLogin']){$[_0xf668('15','V[6Q')]($[_0xf668('16','kmgF')],_0xf668('17','DqQY'),_0xf668('18','CugW')+$[_0xf668('19','xsR)')]+'\x20'+($[_0xf668('1a','rHh*')]||$[_0xf668('1b','dmO7')])+_0xf668('1c','kmgF'),{'open-url':_0xf668('1d','s^UP')});if($[_0xf668('1e','O3X^')]()){await notify[_0xf668('1f','On6g')]($[_0xf668('20','xsR)')]+_0xf668('21','%8LH')+$[_0xf668('22',']l!6')],_0xf668('23','nKU7')+$[_0xf668('24','vPrL')]+'\x20'+$['UserName']+'\x0a请重新登录获取cookie');}continue;}await _0x326225['nXFgh'](entertainment);}}if(_0x326225[_0xf668('25','[#Wy')](message,'')){if(_0x326225[_0xf668('26','x1Fm')]===_0x326225['NcYjL']){if($[_0xf668('27','AA@b')]()){if(_0x326225[_0xf668('28','W6#w')](_0x326225['VRpbB'],_0x326225['KRBnA'])){console[_0xf668('29','AA@b')](e);}else{await notify[_0xf668('2a','AA@b')]($['name']+_0xf668('2b','Ul$D'),message);}}else{await $[_0xf668('2c',']l!6')]($['name']+_0xf668('2d','vPrL'),_0xf668('2e','$eJr'),message);}}else{$[_0xf668('2f','DLvm')](_0x326225[_0xf668('30','922l')]);}}})()[_0xf668('31','@SZ@')](_0x2f70e6=>{$[_0xf668('32','[#Wy')]('','❌\x20'+$['name']+_0xf668('33','#IEW')+_0x2f70e6+'!','');})[_0xf668('34','%8LH')](()=>{$['done']();});async function entertainment(){var _0x5ba172={'Qwotw':function(_0x2fb77b,_0x5f039d){return _0x2fb77b===_0x5f039d;},'FanKK':'WhBiD','infaR':function(_0x32d3c9,_0x376a19){return _0x32d3c9!==_0x376a19;},'HbLUi':_0xf668('35','[XoU'),'XWWEZ':function(_0x4a5f36){return _0x4a5f36();},'BzQMr':function(_0x3b125b,_0x2a6159){return _0x3b125b!==_0x2a6159;},'AzaqG':_0xf668('36','pW0f'),'yPmUz':'headers','wuibq':_0xf668('37','hF7g'),'pqtBR':function(_0x3f932d,_0x34f862){return _0x3f932d===_0x34f862;},'fPKaM':_0xf668('38','KEOw'),'iDviF':_0xf668('39','zzNq'),'CWEKd':_0xf668('3a','V[6Q'),'CSbru':'gzip,\x20deflate,\x20br','DTGtZ':_0xf668('3b','C@fO'),'LXyiy':_0xf668('3c','Ul$D'),'sluWM':'https://api.r2ray.com/jd.bargain/done','BPUUz':'application/json','iHBjq':function(_0x24e239,_0x5d203c){return _0x24e239(_0x5d203c);},'yKaDU':function(_0x2f97a2){return _0x2f97a2();},'vBLSt':function(_0x41e63e){return _0x41e63e();},'ttXli':_0xf668('3d','R[HN'),'oKyZJ':function(_0x3fc77d){return _0x3fc77d();},'rjkiU':function(_0x5e71d1,_0x301d4d){return _0x5e71d1(_0x301d4d);},'BKHWN':'Dwral','pOLfG':'bIJCV','JUknk':_0xf668('3e','vPrL'),'mRABr':_0xf668('3f','%W95'),'gfgQJ':function(_0x4e99b4){return _0x4e99b4();},'lisYV':_0xf668('40','rHh*'),'dmFoL':_0xf668('41','V[6Q'),'Jefbj':function(_0x56b669,_0x44a198){return _0x56b669<_0x44a198;},'iMjMW':function(_0x2869a3,_0x37fa7f){return _0x2869a3!==_0x37fa7f;},'HGtDF':_0xf668('42','C@fO'),'xCpAk':function(_0x1727f6,_0x471d96,_0x4b7464){return _0x1727f6(_0x471d96,_0x4b7464);},'rdzbO':function(_0x2754cc,_0x31a807){return _0x2754cc===_0x31a807;},'WFIJE':function(_0x192819,_0x426501,_0x35085d){return _0x192819(_0x426501,_0x35085d);}};$['risk']=![];$['gameScore']=0x0;$[_0xf668('43','t1Hx')]=0x0;await grantToken();await $[_0xf668('44','wpRu')](0x5dc);await getActCookie();await $[_0xf668('45','uhqA')](0x5dc);await _0x5ba172[_0xf668('46','CugW')](getActInfo);await $[_0xf668('47','O3X^')](0x5dc);await _0x5ba172[_0xf668('48','V[6Q')](getMyPing);await $[_0xf668('49','C@fO')](0x5dc);await _0x5ba172[_0xf668('4a',']AF0')](getUserInfo);await $[_0xf668('4b','yzvA')](0x5dc);await getActContent(![],$[_0xf668('4c','BI4O')]);await doTask(_0xf668('4d','DqQY'),_0xf668('4e','%8LH')+ACT_ID+'&pin='+_0x5ba172[_0xf668('4f','vPrL')](encodeURIComponent,$[_0xf668('50','R[HN')]));if(!$['risk']){var _0x50c305=_0x5ba172[_0xf668('51','BI4O')][_0xf668('52','hF7g')]('|'),_0xf7eca1=0x0;while(!![]){switch(_0x50c305[_0xf7eca1++]){case'0':await answer();continue;case'1':console['log'](_0xf668('53','nKU7')+$['shareCode']+'\x20】');continue;case'2':await _0x5ba172[_0xf668('54','AA@b')](draw);continue;case'3':await $[_0xf668('55','#IEW')](0x5dc);continue;case'4':await getActContent($[_0xf668('56','6BUo')]);continue;case'5':await $[_0xf668('57','QD9b')](0x5dc);continue;case'6':await submitShareCode({'share_code':$[_0xf668('58','2BEr')],'pt_key':$[_0xf668('59','O3X^')]});continue;case'7':await _0x5ba172[_0xf668('5a',']AF0')](getActContent,![]);continue;case'8':await $[_0xf668('5b','xsR)')](0x5dc);continue;}break;}}else{if($[_0xf668('5c','lVb(')]()){if(_0x5ba172[_0xf668('5d','pW0f')]!==_0x5ba172['pOLfG']){await notify[_0xf668('5e','CugW')]($[_0xf668('5f','%W95')]+'运行完成',_0xf668('60','922l')+$['index']+'\x20'+($['nickName']||$[_0xf668('61','%W95')])+_0xf668('62','hF7g'));}else{console['log'](''+JSON['stringify'](err));}}else{if(_0x5ba172['BzQMr'](_0x5ba172[_0xf668('63','C@fO')],_0x5ba172[_0xf668('64','yr@1')])){console['log'](''+JSON[_0xf668('65','[#Wy')](err));}else{await $[_0xf668('66','DLvm')]($['name']+'运行完成',_0xf668('67','^p1!'));}}}if($[_0xf668('68','LgTp')]>0x0){if(_0x5ba172[_0xf668('69','DqQY')](_0x5ba172[_0xf668('6a','C@fO')],_0xf668('6b','uhqA'))){if(data){$['zData']=JSON['parse'](data);};}else{message+=_0xf668('6c','xsR)')+$[_0xf668('19','xsR)')]+'\x20'+($['nickName']||$['UserName'])+_0xf668('6d','vPrL')+$[_0xf668('6e','922l')]+'京豆';}}if(helpAuthor){function _0x3101b3(){var _0x380016={'LMhcn':function(_0x341693,_0x314e1a){return _0x5ba172[_0xf668('6f','xsR)')](_0x341693,_0x314e1a);},'wQtWb':function(_0x1d796c,_0x4a9fa9){return _0x1d796c===_0x4a9fa9;},'htbdE':_0x5ba172['HbLUi'],'IpMWu':function(_0x28eef2){return _0x5ba172[_0xf668('70','ERd%')](_0x28eef2);}};if(_0x5ba172[_0xf668('71','QD9b')](_0x5ba172[_0xf668('72','[#Wy')],_0xf668('73','AA@b'))){$[_0xf668('74','Ul$D')]=data[_0xf668('75','^p1!')][0x0][_0xf668('76','vPrL')];}else{return new Promise(_0x14563b=>{if(_0x5ba172['Qwotw'](_0x5ba172[_0xf668('77','dmO7')],_0xf668('78','t1Hx'))){console[_0xf668('79','pW0f')]('在本地题库中找到了答案：'+tmp[0x0]['a']);options=tmp[0x0]['a'];}else{$['get']({'url':'https://api.r2ray.com/jd.bargain/index'},(_0x5e3a8b,_0x2cbcda,_0x19861c)=>{if(_0x380016[_0xf668('7a','%W95')]('XQyuX',_0xf668('7b','C@fO'))){_0x19861c=JSON[_0xf668('7c','[XoU')](_0x19861c);if(_0x19861c[_0xf668('7d','%W95')]){console[_0xf668('7e','LgTp')](_0xf668('7f','#IEW')+_0x19861c['data'][_0xf668('80','vPrL')]+'】信息获取成功');$['userId']=_0x19861c[_0xf668('81','Ul$D')]['id'];$[_0xf668('82','ERd%')]=_0x19861c[_0xf668('83','nKU7')]['yunMidImageUrl'];$[_0xf668('84','W6#w')]=_0x19861c[_0xf668('85','R[HN')]['nickame'];}else{console[_0xf668('86','DqQY')](_0x19861c);}}else{try{if(_0x19861c){if(_0x380016[_0xf668('87','W6#w')](_0x380016[_0xf668('88',']AF0')],_0x380016[_0xf668('89','V[6Q')])){$[_0xf668('8a','AA@b')]=JSON[_0xf668('8b','@SZ@')](_0x19861c);}else{if(_0x19861c){$[_0xf668('8c','[XoU')]=JSON[_0xf668('8d','9qTx')](_0x19861c);_0x14563b();};}};}catch(_0x452100){console[_0xf668('8e','xsR)')](_0x452100);}finally{_0x380016[_0xf668('8f','ERd%')](_0x14563b);};}});}});}}function _0x42f797(_0x11e7ac,_0x309be){var _0x24bb3a={'cGPnd':_0x5ba172[_0xf668('90','R[HN')],'WZqMW':_0x5ba172[_0xf668('91','WfWy')],'jieyV':function(_0x3c392b,_0x24c11a){return _0x5ba172[_0xf668('92','UDkl')](_0x3c392b,_0x24c11a);},'dYGxI':_0xf668('93','yzvA'),'hrcVI':function(_0x1c1a0a){return _0x1c1a0a();}};let _0x2353bf={'url':_0xf668('94','$eJr'),'headers':{'Host':_0x5ba172[_0xf668('95','x1Fm')],'Content-Type':_0x5ba172[_0xf668('96','V[6Q')],'Origin':_0x5ba172[_0xf668('97','AA@b')],'Accept-Encoding':_0x5ba172['CSbru'],'Cookie':cookie,'Connection':_0xf668('98','lVb('),'Accept':_0x5ba172['DTGtZ'],'User-Agent':_0x5ba172[_0xf668('99','vPrL')],'Referer':'https://h5.m.jd.com/babelDiy/Zeus/4ZK4ZpvoSreRB92RRo8bpJAQNoTq/index.html?serveId=wxe30973feca923229&actId='+_0x11e7ac+_0xf668('9a','s^UP'),'Accept-Language':_0xf668('9b','9qTx')},'body':_0xf668('9c','t1Hx')+_0x11e7ac+_0xf668('9d','WfWy')+_0x309be+_0xf668('9e','C@fO')};return new Promise(_0x45e8e3=>{var _0x1c8ce6={'LfpnR':_0x24bb3a[_0xf668('9f','s^UP')],'UARxS':_0x24bb3a[_0xf668('a0','%8LH')],'aBJKX':function(_0x4ba00b,_0x58a27c){return _0x24bb3a['jieyV'](_0x4ba00b,_0x58a27c);},'jQERS':_0x24bb3a[_0xf668('a1','ERd%')],'QUcJG':function(_0xcbdc37){return _0x24bb3a[_0xf668('a2','On6g')](_0xcbdc37);}};$[_0xf668('a3','O3X^')](_0x2353bf,(_0x41e9ea,_0x4ef3c0,_0x1e6493)=>{if(_0x1e6493){if(_0x1c8ce6[_0xf668('a4','lVb(')](_0x1c8ce6[_0xf668('a5','vPrL')],'smgpS')){$[_0xf668('a6','xsR)')]=JSON['parse'](_0x1e6493);_0x1c8ce6['QUcJG'](_0x45e8e3);}else{cookie=cookie+';';if($[_0xf668('a7','yr@1')]())for(let _0x1b4c0c of resp[_0x1c8ce6[_0xf668('a8','V[6Q')]][_0x1c8ce6[_0xf668('a9','[#Wy')]]){cookie=''+cookie+_0x1b4c0c['split'](';')[0x0]+';';}else{for(let _0x3edced of resp['headers'][_0xf668('aa','C@fO')][_0xf668('ab','KEOw')](',')){cookie=''+cookie+_0x3edced[_0xf668('ac','9qTx')](';')[0x0]+';';}}}};});});}function _0x4bce3b(_0x1627c7,_0x627a45){var _0x23abf7={'sEIRD':function(_0x56b7f1){return _0x5ba172['XWWEZ'](_0x56b7f1);}};let _0x27c4f1={'url':_0x5ba172[_0xf668('ad','uhqA')],'headers':{'Content-Type':_0x5ba172[_0xf668('ae','KEOw')]},'body':JSON[_0xf668('af','R[HN')]({'actID':_0x1627c7,'actsID':_0x627a45,'done':0x1})};return new Promise(_0x1110c1=>{$['post'](_0x27c4f1,(_0x218bd8,_0x5d495f,_0xf6f147)=>{_0x23abf7['sEIRD'](_0x1110c1);});});}await _0x5ba172['gfgQJ'](_0x3101b3);if($['zData'][_0xf668('b0','On6g')][_0xf668('b1','yzvA')]!==0x0){if(_0x5ba172[_0xf668('b2','yzvA')](_0x5ba172['lisYV'],_0x5ba172[_0xf668('b3','QD9b')])){$[_0xf668('b4','x1Fm')]+=data[_0xf668('b5',']AF0')][_0xf668('b6','ERd%')];if(_0x5ba172[_0xf668('b7','922l')](data['data'][_0xf668('b8','UDkl')],0x0)){console['log'](_0xf668('b9','lVb('));}}else{for(let _0x5ed59a=0x0;_0x5ba172[_0xf668('ba','SH*2')](_0x5ed59a,$['zData'][_0xf668('bb','V[6Q')][_0xf668('bc','s^UP')]);_0x5ed59a++){if(_0x5ba172['iMjMW'](_0xf668('bd','DqQY'),_0x5ba172[_0xf668('be','@SZ@')])){var _0x41fcca=_0xf668('bf','^p1!')['split']('|'),_0xdd55b4=0x0;while(!![]){switch(_0x41fcca[_0xdd55b4++]){case'0':actID=$[_0xf668('c0','kmgF')][_0xf668('c1','9qTx')][_0x5ed59a][_0xf668('c2','O3X^')];continue;case'1':await _0x5ba172[_0xf668('c3','R[HN')](_0x42f797,actID,actsID);continue;case'2':if($[_0xf668('c4','$eJr')]&&_0x5ba172[_0xf668('c5','uhqA')]($['Res']['status'],0x4)){await _0x5ba172['WFIJE'](_0x4bce3b,actID,actsID);}continue;case'3':await $[_0xf668('c6','CugW')](0x5dc);continue;case'4':actsID=$[_0xf668('c7','yzvA')]['data'][_0x5ed59a]['actsID'];continue;}break;}}else{_0x5ba172[_0xf668('c8','nKU7')](resolve,data);}};}};};}async function draw(){var _0x4d5557={'BZODG':function(_0x2c26e4,_0x5e13b7){return _0x2c26e4<_0x5e13b7;},'lhZsA':function(_0x512822,_0x399581){return _0x512822===_0x399581;},'RARfn':function(_0x578201,_0x4894bd,_0x187040){return _0x578201(_0x4894bd,_0x187040);},'sBgXz':_0xf668('c9','922l'),'uGRZQ':function(_0x4a8fbb,_0x4000fe){return _0x4a8fbb(_0x4000fe);}};for(let _0x3cbd1a=0x0;_0x4d5557[_0xf668('ca','BI4O')](_0x3cbd1a,$[_0xf668('cb','kmgF')]['length']);_0x3cbd1a++){const _0x5e478b=$[_0xf668('cc','s^UP')][_0x3cbd1a];if(_0x4d5557[_0xf668('cd','$eJr')](_0x5e478b[_0xf668('ce','AA@b')],!![])&&_0x5e478b[_0xf668('cf','QD9b')]===![]){console[_0xf668('8e','xsR)')]('开始抽奖');await _0x4d5557['RARfn'](doTask,_0x4d5557[_0xf668('d0','BI4O')],_0xf668('d1','On6g')+ACT_ID+_0xf668('d2','C@fO')+$['shareCode']+_0xf668('d3','[XoU')+_0x4d5557[_0xf668('d4','V[6Q')](encodeURIComponent,$[_0xf668('d5','V[6Q')])+_0xf668('d6','SH*2')+_0x5e478b['uuid']);await $['wait'](0x3e8);}}}async function answer(){var _0x265202={'YGAiD':function(_0x33572e,_0x2f5dcf){return _0x33572e===_0x2f5dcf;},'psVew':_0xf668('d7','6BUo'),'knDcb':function(_0x4182ab,_0x1c6abb){return _0x4182ab===_0x1c6abb;},'CXJhQ':_0xf668('d8','lVb('),'EfOLO':function(_0x36b0e6,_0x26890c,_0x2cdf9b){return _0x36b0e6(_0x26890c,_0x2cdf9b);},'JWvcl':function(_0x517b47,_0x560481){return _0x517b47(_0x560481);},'HBueD':function(_0x139090,_0xb4299d){return _0x139090<_0xb4299d;},'MsPIi':function(_0x550dbf,_0x37a8d6){return _0x550dbf!==_0x37a8d6;},'bouRA':_0xf668('d9','uhqA'),'VZdac':function(_0x4987ff,_0x401a7c){return _0x4987ff<_0x401a7c;},'arjBc':_0xf668('da',']AF0'),'dSWtD':function(_0x3a0a10,_0x3a1cbf){return _0x3a0a10(_0x3a1cbf);},'CXaex':function(_0x174840,_0x3d11eb,_0x513160){return _0x174840(_0x3d11eb,_0x513160);},'iSyJs':'dingzhi/change/able/answer'};await _0x265202[_0xf668('db','[#Wy')](doTask,_0xf668('dc','On6g'),_0xf668('dd','[XoU')+ACT_ID+'&actorUuid='+$[_0xf668('de','lVb(')]+_0xf668('df','%8LH')+_0x265202['JWvcl'](encodeURIComponent,$[_0xf668('e0','uhqA')]));let _0x26796e=[0x0,0x1,0x2,0x3,0x4,0x5,0x6,0x7,0x8];let _0x2e000a=[];for(let _0x5bc25d=0x0;_0x265202[_0xf668('e1','dmO7')](_0x5bc25d,$[_0xf668('e2',']l!6')][_0xf668('e3','dmO7')]);_0x5bc25d++){const _0x47f2ad=$['cardList'][_0x5bc25d];if(_0x265202[_0xf668('e4','s^UP')](_0x47f2ad[_0xf668('e5','[XoU')],0x63)){if(_0x265202[_0xf668('e6','wpRu')](_0xf668('e7','#IEW'),_0x265202['bouRA'])){_0x2e000a[_0xf668('e8','V[6Q')](_0x47f2ad);}else{$[_0xf668('e9','%8LH')]=data['data'][_0xf668('ea','#IEW')]['baseInfo'][_0xf668('eb','V[6Q')];}}if(_0x265202['MsPIi'](_0x47f2ad['position'],0x63)){let _0x453120=_0x26796e[_0xf668('ec','922l')](_0x47f2ad[_0xf668('ed','DLvm')]);_0x26796e[_0xf668('ee','#IEW')](_0x453120,0x1);}}if(_0x2e000a[_0xf668('ef','CugW')]===0x0){if(_0x265202[_0xf668('f0','%8LH')]('TzgXy','UoEHg')){console[_0xf668('f1','W6#w')](_0xf668('f2','AA@b'));return;}else{$[_0xf668('f3',']AF0')](e,resp);}}if(_0x265202[_0xf668('f4',']l!6')](_0x2e000a['length'],$[_0xf668('f5','922l')])){times=_0x2e000a[_0xf668('f6','WfWy')];}else{times=$['gameScore'];}for(let _0x315f1f=0x0;_0x315f1f<=times;_0x315f1f++){let _0x34657a='';const _0xc90b01=questionList[_0xf668('f7','^p1!')](_0x156478=>_0x156478['q']===_0x2e000a[_0x315f1f][_0xf668('f8','[#Wy')]);if(_0xc90b01&&_0xc90b01[0x0]){if(_0x265202['arjBc']!==_0x265202[_0xf668('f9','pW0f')]){data=JSON['parse'](data);if(_0x265202[_0xf668('fa','AA@b')](data[_0xf668('fb','^p1!')],_0x265202[_0xf668('fc','$eJr')])){$[_0xf668('fd','ERd%')]=![];return;}if(_0x265202[_0xf668('fe','%W95')](data[_0xf668('ff','hF7g')],'0')&&data[_0xf668('b5',']AF0')][_0xf668('100','nKU7')](_0x265202[_0xf668('101','#IEW')])){$['nickName']=data['data'][_0xf668('102',']AF0')][_0xf668('103',']AF0')][_0xf668('104','hF7g')];}}else{console[_0xf668('105','QD9b')](_0xf668('106','On6g')+_0xc90b01[0x0]['a']);_0x34657a=_0xc90b01[0x0]['a'];}}let _0x111bac=_0xf668('107','C@fO')+ACT_ID+_0xf668('108','xsR)')+$[_0xf668('109','KEOw')]+_0xf668('10a',']AF0')+_0x265202[_0xf668('10b','uhqA')](encodeURIComponent,$['secretPin'])+'&uuid='+_0x2e000a[_0x315f1f][_0xf668('10c','LgTp')]+'&answer='+_0x265202[_0xf668('10d','BI4O')](encodeURIComponent,_0x34657a)+_0xf668('10e','s^UP')+_0x26796e[_0x315f1f];await _0x265202['CXaex'](doTask,_0x265202['iSyJs'],_0x111bac);await $[_0xf668('10f','LgTp')](0x5dc);}}async function getActContent(_0xeadc17=!![],_0x1229db=''){var _0x462d3f={'ayppY':function(_0xc6527e,_0x35f2a6){return _0xc6527e===_0x35f2a6;},'vJZxo':_0xf668('110',']l!6'),'NHtBw':_0xf668('111','%8LH'),'yaQnw':function(_0x3f1154,_0x50fabf,_0x3213d5){return _0x3f1154(_0x50fabf,_0x3213d5);},'WpwgQ':function(_0x5b2e67,_0x561b01){return _0x5b2e67(_0x561b01);},'ngAVA':_0xf668('112','%8LH'),'GgbHY':_0xf668('113','UDkl'),'hjkSv':_0xf668('114','rHh*'),'EzekM':function(_0x3f7644,_0x4d96c0){return _0x3f7644+_0x4d96c0;},'WBwPk':_0xf668('115','LgTp'),'qzQTN':function(_0x16d797,_0x8a386){return _0x16d797!==_0x8a386;},'doPiK':_0xf668('116','[#Wy'),'MwOvi':_0xf668('117','uhqA'),'PUxhw':_0xf668('118','wpRu'),'YevxH':_0xf668('119','KEOw')};return new Promise(_0x4b6381=>{var _0x28ff44={'sQCxo':function(_0x37854e){return _0x37854e();},'TSwPJ':function(_0x5332a3,_0x58f409){return _0x462d3f[_0xf668('11a','%8LH')](_0x5332a3,_0x58f409);},'GRRKd':function(_0x26165a,_0x305e7a){return _0x26165a===_0x305e7a;},'nQNmA':_0x462d3f[_0xf668('11b','SH*2')],'AHrgi':function(_0x259fe5,_0x6d263f){return _0x259fe5===_0x6d263f;},'yAuoM':_0x462d3f[_0xf668('11c','yzvA')],'QBLKF':function(_0x357358,_0xded8dd,_0x51bb69){return _0x462d3f[_0xf668('11d','V[6Q')](_0x357358,_0xded8dd,_0x51bb69);},'FJiRD':function(_0x28ec05,_0x3b2d58){return _0x462d3f[_0xf668('11e','QD9b')](_0x28ec05,_0x3b2d58);},'QFknz':function(_0xd2866d,_0x25d018){return _0xd2866d!==_0x25d018;},'KUyOf':_0x462d3f[_0xf668('11f','W6#w')],'pOuzF':_0x462d3f['GgbHY'],'khaOr':_0xf668('120','O3X^'),'sQeHW':_0x462d3f['hjkSv'],'rYMuZ':function(_0x16b378,_0x1ff1ab){return _0x16b378(_0x1ff1ab);},'BTFNd':function(_0x45adbc,_0x5ea3a4){return _0x462d3f['EzekM'](_0x45adbc,_0x5ea3a4);},'rmzAH':_0x462d3f[_0xf668('121',']l!6')],'RFeTI':function(_0x28ca83,_0x2b2d2c){return _0x462d3f[_0xf668('122','kmgF')](_0x28ca83,_0x2b2d2c);},'QOJZQ':_0x462d3f[_0xf668('123','t1Hx')],'xUfNr':function(_0x5e2a30,_0x23c7d1,_0x54981d){return _0x462d3f[_0xf668('124','s^UP')](_0x5e2a30,_0x23c7d1,_0x54981d);},'Gwtxo':_0x462d3f[_0xf668('125','vPrL')],'zMFsG':function(_0x333392,_0x3c1150){return _0x462d3f[_0xf668('126','dmO7')](_0x333392,_0x3c1150);},'GLbwt':function(_0x55e6eb,_0x41b442){return _0x55e6eb===_0x41b442;}};if(_0xf668('127','zzNq')===_0x462d3f['PUxhw']){$[_0xf668('128','@SZ@')](_0x462d3f[_0xf668('129','#IEW')](taskPostUrl,_0x462d3f['YevxH'],_0xf668('12a','uhqA')+ACT_ID+_0xf668('12b','BI4O')+_0x462d3f[_0xf668('12c','hF7g')](encodeURIComponent,$['secretPin'])+_0xf668('12d','DLvm')+$[_0xf668('12e','LgTp')]+'&nick='+$[_0xf668('12f','LgTp')]+_0xf668('130','yzvA')+_0x1229db),async(_0x4a9d47,_0xe2033,_0x2502f2)=>{try{if(_0x4a9d47){console[_0xf668('131','WfWy')](''+JSON[_0xf668('132','QD9b')](_0x4a9d47));}else{_0x2502f2=JSON[_0xf668('133','rHh*')](_0x2502f2);if(_0x2502f2['result']===![]){$[_0xf668('134','zzNq')]=!![];console[_0xf668('135','KEOw')](_0xf668('136','hF7g'));return;}$[_0xf668('137','BI4O')]=_0x2502f2[_0xf668('138','vPrL')]['cardScore'];$[_0xf668('139','R[HN')]=_0x2502f2[_0xf668('13a','2BEr')]['actorUuid'];$[_0xf668('13b','pW0f')]=_0x2502f2['data'][_0xf668('13c','$eJr')];$[_0xf668('13d','%8LH')]=_0x2502f2[_0xf668('13e','DqQY')][_0xf668('13f','LgTp')];$['toShop']=_0x2502f2[_0xf668('c1','9qTx')][_0xf668('140','s^UP')];if(_0x28ff44[_0xf668('141','t1Hx')](_0x2502f2[_0xf668('142','x1Fm')][_0xf668('143','^p1!')],0x9)){if(_0x28ff44['GRRKd'](_0x28ff44[_0xf668('144','WfWy')],_0x28ff44[_0xf668('145','t1Hx')])){$[_0xf668('146','pW0f')]=![];if(_0x2502f2[_0xf668('147','dmO7')]['drawOrNo']===![]&&_0x28ff44['AHrgi'](_0x2502f2['data']['canDrawBig'],!![])){if(_0x28ff44[_0xf668('148','dmO7')]===_0x28ff44[_0xf668('149','Ul$D')]){console[_0xf668('14a','uhqA')]('开始抽取最终大奖。');await _0x28ff44['QBLKF'](doTask,_0xf668('14b','wpRu'),_0xf668('dd','[XoU')+ACT_ID+'&actorUuid='+$[_0xf668('58','2BEr')]+'&pin='+_0x28ff44[_0xf668('14c','C@fO')](escape,$['secretPin'])+'&cardId=');}else{console[_0xf668('14d','yr@1')](e);}}}else{message+=_0xf668('14e','x1Fm')+$[_0xf668('14f','uhqA')]+'\x20'+($[_0xf668('84','W6#w')]||$[_0xf668('150','ERd%')])+_0xf668('151','CugW')+_0x2502f2['data'][_0xf668('152','kmgF')][_0xf668('153','On6g')];}}if(_0xeadc17){if(_0x28ff44[_0xf668('154','CugW')](_0x28ff44[_0xf668('155','CugW')],_0x28ff44['pOuzF'])){for(let _0xec730 of[_0x28ff44[_0xf668('156','O3X^')],_0xf668('157','@SZ@')]){if(_0x28ff44[_0xf668('158','^p1!')]===_0xf668('159','Ul$D')){let _0x2cd231=_0x2502f2[_0xf668('138','vPrL')][_0xec730];for(let _0x51a79d of _0x2cd231['settings']){let _0xc8e706='activityId='+ACT_ID+_0xf668('15a','x1Fm')+$['shareCode']+_0xf668('15b','dmO7')+_0x28ff44['rYMuZ'](encodeURIComponent,$['secretPin'])+_0xf668('15c','yzvA')+_0x51a79d[_0xf668('15d','yzvA')]+'&taskValue='+_0x51a79d[_0xf668('15e','%8LH')];let _0x3dd2ad=_0xf668('15f','2BEr')+_0x2502f2[_0xf668('160','QD9b')][_0xf668('161','@SZ@')]+'&elementId='+encodeURIComponent(_0x28ff44[_0xf668('162','[#Wy')]('店铺',_0x51a79d[_0xf668('163','LgTp')]))+'&pageId='+ACT_ID+_0xf668('164','2BEr')+encodeURIComponent($[_0xf668('165','s^UP')]);if(_0x51a79d[_0xf668('166','[XoU')]===0xc&&_0x28ff44[_0xf668('167','rHh*')](_0x51a79d['status'],0x0)){console[_0xf668('168',']l!6')](_0xf668('169',']l!6')+_0x51a79d[_0xf668('16a','hF7g')]);await doTask('dingzhi/change/able/saveTask',_0xc8e706);await $[_0xf668('16b','UDkl')](0x7d0);await doTask(_0x28ff44[_0xf668('16c','O3X^')],_0x3dd2ad);}else if(_0x51a79d[_0xf668('16d','W6#w')]===0x0){if(_0x28ff44[_0xf668('16e','kmgF')](_0x28ff44[_0xf668('16f','x1Fm')],_0x28ff44['QOJZQ'])){$[_0xf668('170','#IEW')](e,_0xe2033);}else{console['log']('浏览店铺\x20-\x20'+_0x51a79d['name']);await _0x28ff44[_0xf668('171','#IEW')](doTask,_0xf668('172','yzvA'),_0xc8e706);await $['wait'](0x7d0);}}}await $['wait'](0x5dc);}else{console[_0xf668('173','UDkl')](e);}}await $[_0xf668('174','BI4O')](0x5dc);await _0x28ff44[_0xf668('175','W6#w')](doTask,_0x28ff44[_0xf668('176','UDkl')],'activityId='+ACT_ID+_0xf668('177','$eJr')+$[_0xf668('139','R[HN')]+'&pin='+_0x28ff44[_0xf668('178','%8LH')](encodeURIComponent,$[_0xf668('179','dmO7')])+'&taskType='+$[_0xf668('17a','BI4O')][_0xf668('17b','@SZ@')][0x0][_0xf668('15d','yzvA')]+'&taskValue='+$[_0xf668('17c','#IEW')]['settings'][0x0][_0xf668('17d','kmgF')]);}else{_0x28ff44['sQCxo'](_0x4b6381);}}}}catch(_0x50399b){$[_0xf668('17e','QD9b')](_0x50399b,_0xe2033);}finally{_0x28ff44['zMFsG'](_0x4b6381,_0x2502f2);}});}else{if(err){console['log'](''+JSON[_0xf668('17f','SH*2')](err));}else{data=JSON[_0xf668('180','zzNq')](data);if(_0x28ff44[_0xf668('181','R[HN')](data['code'],'0')){$[_0xf668('182','6BUo')]=data[_0xf668('183','rHh*')];}}}});}function doTask(_0x25c45b,_0x17d2e7){var _0x54e6fd={'pqzAH':function(_0x34d421,_0x2f2705){return _0x34d421===_0x2f2705;},'ChbyM':_0xf668('184','s^UP'),'guKFJ':_0xf668('185','DLvm'),'jBWuM':function(_0x23a2ab,_0x9c19a4){return _0x23a2ab===_0x9c19a4;},'VPWaU':_0xf668('186','vPrL'),'GUwUZ':_0xf668('187','dmO7'),'RXtVd':_0xf668('37','hF7g'),'NmHtN':function(_0x57585f,_0x17fdef){return _0x57585f===_0x17fdef;},'aGtHW':_0xf668('188','9qTx'),'SitDr':function(_0x36c884,_0x20c232){return _0x36c884===_0x20c232;},'mwPhf':_0xf668('189','W6#w'),'BixxQ':function(_0x3c6866,_0x20de21){return _0x3c6866!==_0x20de21;},'AnVAM':_0xf668('18a','QD9b'),'PaOVV':function(_0x583c75,_0x3044f3){return _0x583c75!==_0x3044f3;},'ilOxT':_0xf668('18b','ERd%'),'IOGqP':_0xf668('18c','V[6Q'),'agccT':function(_0x13f681,_0x461cbb){return _0x13f681===_0x461cbb;},'MIiAk':function(_0x588198,_0x33c6bf){return _0x588198!==_0x33c6bf;},'xAmhp':'buvju','QWRES':_0xf668('18d','[#Wy'),'QMCaG':_0xf668('18e','lVb('),'EsGAq':_0xf668('18f','V[6Q'),'rPali':function(_0x18b228){return _0x18b228();},'NOEHE':_0xf668('190','922l'),'KBqOj':function(_0x342dd2,_0x1fd9c0,_0x30a6a8){return _0x342dd2(_0x1fd9c0,_0x30a6a8);}};return new Promise(_0x220a50=>{var _0x1dfb65={'obLKY':function(_0x522a9b){return _0x54e6fd[_0xf668('191','uhqA')](_0x522a9b);},'NRjRb':_0xf668('192','LgTp')};if(_0x54e6fd[_0xf668('193','9qTx')]!=='AMhkz'){$['post'](_0x54e6fd[_0xf668('194','%8LH')](taskPostUrl,_0x25c45b,_0x17d2e7),(_0x1e79ca,_0x3e8a7d,_0x1a06bc)=>{var _0x1051ce={'PyYSa':function(_0x5d9105,_0x2a3351){return _0x54e6fd[_0xf668('195','O3X^')](_0x5d9105,_0x2a3351);}};if(_0x54e6fd[_0xf668('196','2BEr')]!==_0x54e6fd[_0xf668('197','yr@1')]){try{if(_0x1e79ca){console['log'](''+JSON[_0xf668('af','R[HN')](_0x1e79ca));}else{if(_0x54e6fd[_0xf668('198','t1Hx')](_0x54e6fd[_0xf668('199','UDkl')],_0x54e6fd[_0xf668('19a','pW0f')])){if(_0x1a06bc){_0x1a06bc=JSON[_0xf668('19b','V[6Q')](_0x1a06bc);if(_0x3e8a7d[_0x54e6fd[_0xf668('19c','$eJr')]][_0xf668('19d','%8LH')]){cookie=_0x3e8a7d[_0x54e6fd[_0xf668('19e','922l')]][_0x54e6fd[_0xf668('19f','QD9b')]][_0xf668('1a0','s^UP')](';')+';\x20'+originCookie;}if(_0x54e6fd['jBWuM'](_0x1a06bc['result'],!![])){if(_0x54e6fd['NmHtN'](_0x54e6fd[_0xf668('1a1','zzNq')],_0xf668('1a2','pW0f'))){_0x1dfb65['obLKY'](_0x220a50);}else{if(_0x1a06bc['data']['hasOwnProperty'](_0xf668('1a3','SH*2'))){if(_0x54e6fd[_0xf668('1a4','%W95')](_0x54e6fd[_0xf668('1a5','#IEW')],_0xf668('1a6','WfWy'))){$[_0xf668('1a7','xsR)')]=_0x1a06bc['data'][_0xf668('1a8','2BEr')];}else{$[_0xf668('1a9','t1Hx')]+=_0x1a06bc[_0xf668('1aa','WfWy')][_0xf668('1ab','O3X^')];if(_0x54e6fd[_0xf668('1ac','2BEr')](_0x1a06bc[_0xf668('b5',']AF0')][_0xf668('1ad','WfWy')],0x0)){console[_0xf668('1ae','^p1!')](_0xf668('1af','SH*2'));}}}if(_0x1a06bc[_0xf668('142','x1Fm')][_0xf668('1b0','t1Hx')](_0x54e6fd[_0xf668('1b1','BI4O')])){if(_0x54e6fd['BixxQ']('jZduJ',_0xf668('1b2','SH*2'))){_0x1dfb65[_0xf668('1b3','wpRu')](_0x220a50);}else{$[_0xf668('1b4','$eJr')]=_0x1a06bc[_0xf668('1b5','O3X^')][_0xf668('1b6','CugW')];}}if(_0x1a06bc[_0xf668('b5',']AF0')][_0xf668('1b7','@SZ@')]('right')){if(_0x1a06bc['data']['right']===!![]){console[_0xf668('1b8','C@fO')](_0xf668('1b9','xsR)'));}}if(_0x1a06bc[_0xf668('1ba','SH*2')][_0xf668('1bb','9qTx')]('drawInfo')){if(_0x54e6fd[_0xf668('1bc','SH*2')](_0x54e6fd[_0xf668('1bd','922l')],_0x54e6fd[_0xf668('1be','zzNq')])){if(_0x54e6fd[_0xf668('1bf','pW0f')](_0x1a06bc[_0xf668('13e','DqQY')][_0xf668('1c0','xsR)')],!![])){message+='获得'+_0x1a06bc['data'][_0xf668('1c1','$eJr')][_0xf668('20','xsR)')]+'\x0a';console['log']('获得'+_0x1a06bc[_0xf668('83','nKU7')][_0xf668('1c2','On6g')][_0xf668('153','On6g')]+'\x0a');if(_0x1a06bc[_0xf668('1c3','%8LH')][_0xf668('1c4','DqQY')]===0x6){$['bean']+=_0x1a06bc[_0xf668('1c5','C@fO')][_0xf668('1c6','dmO7')]['beanNum'];}else{if(_0x54e6fd[_0xf668('1c7','ERd%')](_0x54e6fd[_0xf668('1c8','[#Wy')],'Zgmur')){message+=_0xf668('1c9',']AF0')+$[_0xf668('1ca','WfWy')]+'\x20'+($[_0xf668('1cb','t1Hx')]||$['UserName'])+_0xf668('1cc','lVb(')+_0x1a06bc[_0xf668('7d','%W95')]['drawInfo'][_0xf668('1cd','WfWy')];}else{console['log'](e,_0x3e8a7d);}}}else{console[_0xf668('1ce','zzNq')](_0xf668('1cf','%8LH')+_0x1a06bc[_0xf668('1d0','BI4O')]);}}else{cookie=''+cookie+ck[_0xf668('1d1','t1Hx')](';')[0x0]+';';}}}}else{console[_0xf668('1d2','[XoU')](_0x1a06bc[_0xf668('1d3','kmgF')]);}}else{if(_0x54e6fd['agccT'](_0x54e6fd[_0xf668('1d4','vPrL')],_0xf668('1d5','wpRu'))){$[_0xf668('1d6','s^UP')](_0x54e6fd[_0xf668('1d7','WfWy')]);}else{_0x1a06bc=JSON['parse'](_0x1a06bc);if(_0x1a06bc['result']){$[_0xf668('1d8','zzNq')]=_0x1a06bc['data'][_0xf668('1d9','BI4O')];}}}}else{if(_0x1e79ca){console[_0xf668('1da','x1Fm')](''+JSON['stringify'](_0x1e79ca));}else{_0x1a06bc=JSON[_0xf668('180','zzNq')](_0x1a06bc);if(_0x1051ce[_0xf668('1db','zzNq')](_0x1a06bc[_0xf668('1dc','@SZ@')],'0')){$[_0xf668('1dd','DLvm')]=_0x1a06bc['tokenKey'];cookie=cookie+'IsvToken='+$[_0xf668('1de','2BEr')];}}}}}catch(_0x37a38d){if(_0x54e6fd[_0xf668('1df','lVb(')](_0x54e6fd['EsGAq'],'QiVpm')){console[_0xf668('14d','yr@1')](_0x37a38d,_0x3e8a7d);}else{_0x220a50(_0x1a06bc);}}finally{_0x220a50();}}else{$[_0xf668('1e0','lVb(')](_0x1dfb65['NRjRb']);}});}else{$[_0xf668('183','rHh*')]=data[_0xf668('1e1','On6g')];}});}function getUserInfo(){var _0x4eb35a={'Xqpch':'vULLV','QorwS':function(_0x249938,_0x246961){return _0x249938!==_0x246961;},'DpqzB':_0xf668('1e2','[XoU'),'Wjvvv':'VWRDz','qqUmu':function(_0x438f31,_0x2f4a69){return _0x438f31(_0x2f4a69);},'sImRE':function(_0x5a1a27,_0x46e191){return _0x5a1a27===_0x46e191;},'NYzuk':_0xf668('1e3','9qTx'),'UhQUM':_0xf668('1e4','hF7g'),'pZUbu':function(_0x123665,_0x370a74,_0x292ba6){return _0x123665(_0x370a74,_0x292ba6);},'dyjMG':'wxActionCommon/getUserInfo'};return new Promise(_0x50b846=>{var _0x39bc52={'MHpQj':function(_0x3ded85,_0x46762c){return _0x4eb35a['sImRE'](_0x3ded85,_0x46762c);},'YWupb':_0x4eb35a[_0xf668('1e5','s^UP')],'YwCWo':_0x4eb35a['UhQUM']};let _0x505af3='pin='+_0x4eb35a[_0xf668('1e6','LgTp')](encodeURIComponent,$['secretPin']);$[_0xf668('1e7','lVb(')](_0x4eb35a[_0xf668('1e8','QD9b')](taskPostUrl,_0x4eb35a['dyjMG'],_0x505af3),async(_0x3d71f1,_0x3cfc2b,_0x301402)=>{var _0x358c76={'HZFib':function(_0x145c1b,_0x57a9aa){return _0x145c1b(_0x57a9aa);}};if(_0x4eb35a['Xqpch']===_0x4eb35a['Xqpch']){try{if(_0x4eb35a['QorwS'](_0x4eb35a[_0xf668('1e9','%W95')],_0xf668('1ea','zzNq'))){if(_0x3d71f1){console[_0xf668('1eb','%W95')](''+JSON['stringify'](_0x3d71f1));}else{_0x301402=JSON[_0xf668('8d','9qTx')](_0x301402);if(_0x301402['data']){if(_0xf668('1ec','$eJr')===_0x4eb35a[_0xf668('1ed','lVb(')]){console[_0xf668('1ee','6BUo')](_0xf668('1ef','[#Wy')+_0x301402[_0xf668('1f0','#IEW')][_0xf668('1f1','s^UP')]+_0xf668('1f2','DqQY'));$[_0xf668('1f3','922l')]=_0x301402['data']['id'];$[_0xf668('1f4','kmgF')]=_0x301402['data'][_0xf668('1f5','CugW')];$[_0xf668('1f6','^p1!')]=_0x301402[_0xf668('147','dmO7')][_0xf668('1f7','DqQY')];}else{_0x358c76[_0xf668('1f8','C@fO')](_0x50b846,_0x301402);}}else{console[_0xf668('168',']l!6')](_0x301402);}}}else{_0x301402=JSON[_0xf668('8d','9qTx')](_0x301402);if(_0x39bc52[_0xf668('1f9','9qTx')](_0x301402['code'],'0')){$[_0xf668('1fa','V[6Q')]=_0x301402[_0xf668('182','6BUo')];}}}catch(_0x22c1a6){$[_0xf668('1fb','R[HN')](_0x22c1a6,_0x3cfc2b);}finally{_0x4eb35a['qqUmu'](_0x50b846,_0x301402);}}else{var _0x17462f={'enVjR':function(_0x27c6be){return _0x27c6be();}};let _0x5e412a={'url':_0x39bc52['YWupb'],'headers':{'Content-Type':_0x39bc52['YwCWo']},'body':JSON[_0xf668('1fc','AA@b')]({'actID':actID,'actsID':actsID,'done':0x1})};return new Promise(_0x302cde=>{$[_0xf668('1fd','dmO7')](_0x5e412a,(_0x1b339b,_0x3239f5,_0x240fd6)=>{_0x17462f['enVjR'](_0x302cde);});});}});});}function getMyPing(){var _0x4231c5={'btOpK':function(_0x43032a,_0x263c3b){return _0x43032a!==_0x263c3b;},'wcktW':'tzovy','WnmjP':function(_0x4000d0,_0x428ba8,_0x3ccffc){return _0x4000d0(_0x428ba8,_0x3ccffc);},'JvNAL':'customer/getMyPing'};return new Promise(_0x3d9d47=>{var _0x1be81d={'XtBxB':function(_0x24c9f9,_0x94a3c5){return _0x24c9f9>_0x94a3c5;},'wzwEm':function(_0xff1e66,_0x5ac644){return _0x4231c5[_0xf668('1fe','nKU7')](_0xff1e66,_0x5ac644);},'tWwhf':_0x4231c5[_0xf668('1ff','KEOw')],'YXZPs':_0xf668('200','UDkl')};$[_0xf668('201','DLvm')](_0x4231c5[_0xf668('202','C@fO')](taskPostUrl,_0x4231c5[_0xf668('203','x1Fm')],_0xf668('204','UDkl')+$[_0xf668('205','x1Fm')]+_0xf668('206','dmO7')+$[_0xf668('207','kmgF')]+_0xf668('208','BI4O')),async(_0x31e88c,_0x339ca1,_0x1211c2)=>{var _0x44ae1e={'JJmez':function(_0x3fcd3b,_0x1a78db){return _0x1be81d['XtBxB'](_0x3fcd3b,_0x1a78db);}};try{if(_0x1be81d[_0xf668('209','$eJr')](_0xf668('20a','BI4O'),_0x1be81d[_0xf668('20b','DLvm')])){times=newCardList['length'];}else{if(_0x31e88c){console[_0xf668('20c',']AF0')](''+JSON[_0xf668('20d','t1Hx')](_0x31e88c));}else{if(_0x1be81d[_0xf668('20e','[XoU')](_0x1be81d[_0xf668('20f',']AF0')],_0xf668('210','SH*2'))){_0x1211c2=JSON[_0xf668('8d','9qTx')](_0x1211c2);if(_0x1211c2[_0xf668('211','DqQY')]){$['secretPin']=_0x1211c2[_0xf668('85','R[HN')][_0xf668('d5','V[6Q')];cookie=_0xf668('212','dmO7')+$[_0xf668('50','R[HN')]+';'+cookie;}}else{if(_0x1211c2){_0x1211c2=JSON[_0xf668('213','t1Hx')](_0x1211c2);if(_0x44ae1e[_0xf668('214','%W95')](_0x1211c2[_0xf668('215','ERd%')]['length'],0x0)){$[_0xf668('216','dmO7')]=_0x1211c2['data'][0x0][_0xf668('217','SH*2')];}else{$[_0xf668('218','SH*2')]='';}}}}}}catch(_0x31400e){$['logErr'](_0x31400e,_0x339ca1);}finally{_0x3d9d47();}});});}function getActInfo(){var _0x54c0c3={'mgMjy':function(_0x4b0f73,_0x22536b){return _0x4b0f73===_0x22536b;},'TIBxo':function(_0x59e446,_0x3831b8){return _0x59e446!==_0x3831b8;},'YhvRg':function(_0x75ace8,_0x1a6db5){return _0x75ace8(_0x1a6db5);},'WExst':function(_0x557d18,_0x21d51a){return _0x557d18!==_0x21d51a;},'GRAQy':_0xf668('219','2BEr'),'UUEbB':'dykBf','sLpsF':_0xf668('21a','6BUo'),'gTiwp':'pkxwQ','uWDYp':_0xf668('21b','WfWy')};return new Promise(_0x3bde45=>{if('GsdGu'===_0x54c0c3['gTiwp']){data=JSON[_0xf668('21c','$eJr')](data);if(_0x54c0c3[_0xf668('21d','On6g')](data['code'],'0')){$[_0xf668('21e','vPrL')]=data[_0xf668('21f','LgTp')];cookie=cookie+'IsvToken='+$['tokenKey'];}}else{$[_0xf668('220','R[HN')](taskPostUrl(_0x54c0c3['uWDYp'],'activityId='+ACT_ID),async(_0x4502f8,_0x3c236c,_0x2698b3)=>{var _0x230472={'CzzFb':function(_0x20a91a,_0x2c503b){return _0x20a91a===_0x2c503b;},'HLPVQ':function(_0x4de065,_0x4e4fd6){return _0x54c0c3[_0xf668('221','KEOw')](_0x4de065,_0x4e4fd6);},'JyoHN':function(_0x367330,_0x3bfc88){return _0x54c0c3[_0xf668('222','lVb(')](_0x367330,_0x3bfc88);}};if(_0x54c0c3[_0xf668('223','2BEr')](_0xf668('224','9qTx'),_0xf668('225','hF7g'))){if(_0x4502f8){console[_0xf668('226','9qTx')](''+JSON[_0xf668('227','wpRu')](_0x4502f8));}else{_0x2698b3=JSON['parse'](_0x2698b3);if(_0x2698b3[_0xf668('228','#IEW')]){$[_0xf668('229','LgTp')]=_0x2698b3[_0xf668('22a','UDkl')][_0xf668('22b','rHh*')];cookie='AUTH_C_USER='+$[_0xf668('22c','ERd%')]+';'+cookie;}}}else{try{if(_0x4502f8){if(_0x54c0c3[_0xf668('22d','[#Wy')]!==_0x54c0c3[_0xf668('22e','@SZ@')]){console[_0xf668('1ce','zzNq')](''+JSON[_0xf668('22f','kmgF')](_0x4502f8));}else{console[_0xf668('32','[#Wy')](''+JSON[_0xf668('17f','SH*2')](_0x4502f8));}}else{_0x2698b3=JSON[_0xf668('230','R[HN')](_0x2698b3);if(_0x2698b3[_0xf668('231','%W95')]){$[_0xf668('232','t1Hx')]=_0x2698b3[_0xf668('233','KEOw')][_0xf668('234','xsR)')];}}}catch(_0x347dd0){if('dykBf'===_0x54c0c3[_0xf668('235','DLvm')]){$['logErr'](_0x347dd0,_0x3c236c);}else{const _0x316b7a=$[_0xf668('236','wpRu')][i];if(_0x230472['CzzFb'](_0x316b7a[_0xf668('e5','[XoU')],0x63)){newCardList[_0xf668('237','9qTx')](_0x316b7a);}if(_0x230472[_0xf668('238','s^UP')](_0x316b7a['position'],0x63)){let _0x1f5b18=newPosition[_0xf668('239',']AF0')](_0x316b7a[_0xf668('23a','UDkl')]);newPosition['splice'](_0x1f5b18,0x1);}}}finally{if(_0x54c0c3[_0xf668('23b','xsR)')](_0x54c0c3[_0xf668('23c','O3X^')],_0x54c0c3[_0xf668('23d','s^UP')])){_0x54c0c3[_0xf668('23e','xsR)')](_0x3bde45,_0x2698b3);}else{_0x230472[_0xf668('23f','t1Hx')](_0x3bde45,_0x2698b3);}}}});}});}function grantTokenKey(){var _0x325b3d={'QhyNY':function(_0x5a7192,_0x17d88d){return _0x5a7192===_0x17d88d;},'egoBP':_0xf668('240','SH*2'),'ZfBzh':'aCqbL','CCOrB':function(_0x412016,_0x4915b6){return _0x412016!==_0x4915b6;},'UtIxq':'NgDRG','yZbEB':_0xf668('241','@SZ@'),'URcbv':_0xf668('242','yr@1'),'HbTEE':'ydgWM','PtFZD':_0xf668('243','%8LH'),'corZq':function(_0x10fe39){return _0x10fe39();},'lWpRs':_0xf668('244','W6#w'),'TzMKK':_0xf668('245','yr@1'),'pKleZ':_0xf668('246','vPrL'),'GYWOl':_0xf668('247','#IEW'),'ICBti':_0xf668('248','s^UP')};let _0x1c0065={'url':_0xf668('249','V[6Q'),'headers':{'Host':_0x325b3d['lWpRs'],'Content-Type':_0x325b3d[_0xf668('24a','nKU7')],'Accept':_0x325b3d[_0xf668('24b','[#Wy')],'Connection':_0x325b3d['GYWOl'],'Cookie':cookie,'User-Agent':'JD4iPhone/167538\x20(iPhone;\x20iOS\x2014.3;\x20Scale/3.00)','Accept-Language':_0x325b3d[_0xf668('24c','R[HN')],'Accept-Encoding':_0xf668('24d','[#Wy')},'body':_0xf668('24e','lVb(')};return new Promise(_0x2c2765=>{$[_0xf668('24f','CugW')](_0x1c0065,(_0x3ab2c7,_0x33d053,_0x49e37d)=>{try{if(_0x325b3d[_0xf668('250','zzNq')](_0x325b3d['egoBP'],_0x325b3d['ZfBzh'])){$['tokenKey']=_0x49e37d['tokenKey'];cookie=cookie+_0xf668('251','yr@1')+$[_0xf668('252','xsR)')];}else{if(_0x3ab2c7){console[_0xf668('32','[#Wy')](''+JSON[_0xf668('253',']AF0')](_0x3ab2c7));}else{_0x49e37d=JSON[_0xf668('254','W6#w')](_0x49e37d);if(_0x49e37d[_0xf668('1dc','@SZ@')]==='0'){if(_0x325b3d[_0xf668('255','AA@b')](_0x325b3d[_0xf668('256','Ul$D')],_0x325b3d['yZbEB'])){$['tokenKey']=_0x49e37d[_0xf668('257','yr@1')];cookie=cookie+_0xf668('258',']l!6')+$[_0xf668('259','6BUo')];}else{$[_0xf668('25a','dmO7')]=JSON[_0xf668('25b','BI4O')](_0x49e37d);}}}}}catch(_0x1a0daf){if(_0x325b3d[_0xf668('25c','WfWy')](_0x325b3d[_0xf668('25d','R[HN')],_0x325b3d[_0xf668('25e','WfWy')])){_0x49e37d=JSON['parse'](_0x49e37d);if(_0x49e37d['result']){$[_0xf668('165','s^UP')]=_0x49e37d[_0xf668('215','ERd%')][_0xf668('25f','x1Fm')];cookie=_0xf668('260','[#Wy')+$[_0xf668('261','^p1!')]+';'+cookie;}}else{console[_0xf668('262','SH*2')](_0x1a0daf,_0x33d053);}}finally{if(_0x325b3d[_0xf668('263','Ul$D')]===_0x325b3d[_0xf668('264','On6g')]){_0x325b3d[_0xf668('265','BI4O')](_0x2c2765);}else{$[_0xf668('216','dmO7')]='';}}});});}function grantToken(){var _0x77697d={'lIPTD':function(_0x1cbdab){return _0x1cbdab();},'LSBKH':function(_0x3d704f,_0x491311){return _0x3d704f!==_0x491311;},'AwXOG':'wZlWp','XHcgo':_0xf668('266','C@fO'),'aMWJA':function(_0x145bdc,_0x30aed2){return _0x145bdc===_0x30aed2;},'KakJV':'jMQHF','hhWyL':_0xf668('267','SH*2'),'qOLFs':_0xf668('268','uhqA'),'pxpaZ':_0xf668('269','zzNq'),'itIuM':_0xf668('26a','SH*2'),'FNssH':_0xf668('26b','Ul$D'),'bOOjY':'zh-Hans-CN;q=1','vsFns':_0xf668('26c','922l')};let _0x5f2048={'url':_0x77697d[_0xf668('26d','vPrL')],'headers':{'Host':_0xf668('26e','BI4O'),'Content-Type':_0x77697d['itIuM'],'Accept':_0xf668('26f','KEOw'),'Connection':_0x77697d['FNssH'],'Cookie':cookie,'User-Agent':_0xf668('270','t1Hx'),'Accept-Language':_0x77697d[_0xf668('271','^p1!')],'Accept-Encoding':_0x77697d[_0xf668('272','@SZ@')]},'body':_0xf668('273','%W95')};return new Promise(_0x7bb504=>{$[_0xf668('274','#IEW')](_0x5f2048,(_0x57e399,_0x30e509,_0x28be9f)=>{var _0x47aa8b={'JIdCU':function(_0x3f34e9){return _0x77697d[_0xf668('275','%W95')](_0x3f34e9);}};if(_0x77697d[_0xf668('276','UDkl')](_0x77697d['AwXOG'],_0xf668('277','pW0f'))){console[_0xf668('278','CugW')](''+JSON[_0xf668('279','uhqA')](_0x57e399));}else{try{if(_0x57e399){if(_0x77697d[_0xf668('27a','[XoU')](_0xf668('27b','$eJr'),_0x77697d['XHcgo'])){_0x47aa8b[_0xf668('27c','DLvm')](_0x7bb504);}else{console['log'](''+JSON['stringify'](_0x57e399));}}else{_0x28be9f=JSON['parse'](_0x28be9f);if(_0x77697d[_0xf668('27d','DLvm')](_0x28be9f[_0xf668('27e','lVb(')],'0')){if(_0x77697d[_0xf668('27f','t1Hx')](_0x77697d[_0xf668('280','DLvm')],_0x77697d[_0xf668('281','lVb(')])){$[_0xf668('282','[XoU')]=_0x28be9f['token'];}else{$[_0xf668('283','lVb(')]+=_0x28be9f[_0xf668('284','yzvA')][_0xf668('285','AA@b')][_0xf668('286','pW0f')];}}}}catch(_0x3f2c08){if(_0x77697d[_0xf668('287','dmO7')]===_0xf668('288','Ul$D')){console['log'](''+JSON[_0xf668('289','rHh*')](_0x57e399));}else{console[_0xf668('1d6','s^UP')](_0x3f2c08);}}finally{_0x77697d[_0xf668('28a','9qTx')](_0x7bb504);}}});});}function getActCookie(){var _0x58fae8={'pDlAc':_0xf668('28b','V[6Q'),'ncjzJ':function(_0x1971c2,_0x6cb44){return _0x1971c2===_0x6cb44;},'SRnLx':'FHumH','Innfc':_0xf668('28c','Ul$D'),'iheYl':_0xf668('28d','QD9b'),'rCKaE':_0xf668('28e','#IEW'),'VFMIN':'iRbxE','BQSxe':_0xf668('28f','xsR)'),'RQbob':function(_0x2e813c){return _0x2e813c();},'HQKnh':'【提示】请先获取京东账号一cookie\x0a直接使用NobyDa的京东签到获取','tCxNq':'https://bean.m.jd.com/bean/signIndex.action','nHfej':function(_0x17893b,_0x4f1aac){return _0x17893b!==_0x4f1aac;},'deGXp':'IbEXJ','UmDxT':'text/plain','jdCMh':'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8','hypOf':_0xf668('26b','Ul$D'),'FtSta':'jdapp;iPhone;9.3.8;14.3;network/wifi;ADID/;supportApplePay/0;hasUPPay/0;hasOCPay/0;model/iPhone10,3;supportBestPay/0;appBuild/167538;jdSupportDarkMode/0;addressid/0;pv/1.12;apprpd/Babel_Native;ref/JDWebViewController;psq/11;ads/;psn/;jdv/0|;adk/;app_device/IOS;pap/JA2015_311210|9.3.8|IOS\x2014.3;Mozilla/5.0\x20(iPhone;\x20CPU\x20iPhone\x20OS\x2014_3\x20like\x20Mac\x20OS\x20X)\x20AppleWebKit/605.1.15\x20(KHTML,\x20like\x20Gecko)\x20Mobile/15E148;supportJDSHWK/1','YYoLk':'zh-Hans-CN;q=1'};let _0x51a13e={'url':_0xf668('290','DqQY')+ACT_ID,'headers':{'Content-Type':_0x58fae8[_0xf668('291','lVb(')],'Accept':_0x58fae8[_0xf668('292','rHh*')],'Connection':_0x58fae8[_0xf668('293','kmgF')],'Cookie':''+cookie,'User-Agent':_0x58fae8[_0xf668('294','pW0f')],'Accept-Language':_0x58fae8['YYoLk'],'Accept-Encoding':'gzip,\x20deflate,\x20br'}};return new Promise(_0x4b7e17=>{var _0x33297a={'mBRBG':'headers','BSIIR':_0xf668('295','[#Wy')};if(_0x58fae8[_0xf668('296','x1Fm')]('WQopp',_0x58fae8[_0xf668('297','W6#w')])){$[_0xf668('298','C@fO')](_0x51a13e,(_0x1b7f9b,_0x5c4eec,_0x398451)=>{try{if(_0x1b7f9b){if(_0xf668('28b','V[6Q')===_0x58fae8[_0xf668('299','CugW')]){console['log'](''+JSON[_0xf668('29a','6BUo')](_0x1b7f9b));}else{$[_0xf668('29b','s^UP')](e,_0x5c4eec);}}else{if(_0x58fae8[_0xf668('29c','^p1!')](_0x58fae8[_0xf668('29d','9qTx')],_0x58fae8[_0xf668('29e','rHh*')])){times=$[_0xf668('29f','rHh*')];}else{cookie=cookie+';';if($[_0xf668('2a0','UDkl')]())for(let _0x342aad of _0x5c4eec[_0xf668('2a1','x1Fm')][_0xf668('37','hF7g')]){if(_0x58fae8[_0xf668('2a2','O3X^')](_0xf668('2a3','%W95'),_0x58fae8['iheYl'])){for(let _0x5a8837 of _0x5c4eec[_0x33297a[_0xf668('2a4',']AF0')]][_0x33297a[_0xf668('2a5','@SZ@')]][_0xf668('2a6','lVb(')](',')){cookie=''+cookie+_0x5a8837['split'](';')[0x0]+';';}}else{cookie=''+cookie+_0x342aad[_0xf668('2a7','C@fO')](';')[0x0]+';';}}else{for(let _0x10cac8 of _0x5c4eec[_0x58fae8[_0xf668('2a8','yzvA')]]['Set-Cookie'][_0xf668('2a9','%W95')](',')){if(_0x58fae8[_0xf668('2aa','DqQY')](_0x58fae8[_0xf668('2ab','W6#w')],_0x58fae8[_0xf668('2ac','DLvm')])){console[_0xf668('f1','W6#w')](''+JSON[_0xf668('2ad','BI4O')](_0x1b7f9b));}else{cookie=''+cookie+_0x10cac8[_0xf668('2ae','2BEr')](';')[0x0]+';';}}}}}}catch(_0x35f1a5){console[_0xf668('278','CugW')](_0x35f1a5);}finally{_0x58fae8[_0xf668('2af','@SZ@')](_0x4b7e17);}});}else{$[_0xf668('2b0','R[HN')]($['name'],_0x58fae8[_0xf668('2b1','pW0f')],_0x58fae8[_0xf668('2b2','kmgF')],{'open-url':_0x58fae8[_0xf668('2b3','xsR)')]});return;}});}function taskPostUrl(_0xadfcbc,_0x4dd364){var _0x5d4563={'rHmFr':_0xf668('2b4','9qTx'),'fyvdo':_0xf668('2b5','UDkl'),'YBEeH':_0xf668('2b6','BI4O'),'zLLiQ':_0xf668('2b7','uhqA'),'LRnXY':_0xf668('2b8','zzNq'),'pKzlM':_0xf668('2b9','pW0f')};return{'url':_0xf668('2ba','nKU7')+_0xadfcbc,'headers':{'Host':_0x5d4563[_0xf668('2bb','UDkl')],'Accept':'application/json','X-Requested-With':_0xf668('2bc','pW0f'),'Accept-Language':_0x5d4563[_0xf668('2bd','C@fO')],'Accept-Encoding':_0x5d4563[_0xf668('2be','nKU7')],'Content-Type':_0x5d4563['zLLiQ'],'Origin':_0x5d4563[_0xf668('2bf','6BUo')],'Connection':'keep-alive','Referer':_0xf668('2c0','C@fO')+ACT_ID,'Cookie':''+cookie,'User-Agent':_0x5d4563['pKzlM']},'body':_0x4dd364};}function getShareCode(){var _0x51c48c={'yVHhU':function(_0x330677){return _0x330677();},'hFPQa':function(_0x45c5e2,_0x284a04){return _0x45c5e2!==_0x284a04;},'vYFfK':_0xf668('2c1','UDkl'),'ZuHUg':'eHYbO','kJeMt':function(_0x5bc6a9,_0x18a2e5){return _0x5bc6a9>_0x18a2e5;},'KCdyN':'LLuqS','VvobG':_0xf668('2c2','SH*2')};return new Promise(_0x3fa36f=>{var _0x5be915={'uRNnq':function(_0x3ef76d){return _0x51c48c['yVHhU'](_0x3ef76d);},'urgnh':function(_0x1c1470,_0x1d1db7){return _0x51c48c[_0xf668('2c3','[XoU')](_0x1c1470,_0x1d1db7);},'FuWzi':_0x51c48c[_0xf668('2c4','@SZ@')],'xbxni':function(_0x11dfbd,_0x3e0064){return _0x51c48c[_0xf668('2c5','uhqA')](_0x11dfbd,_0x3e0064);},'GffTA':'VciJn','tXFHX':_0x51c48c['ZuHUg'],'MNzjQ':function(_0x205411,_0xf66968){return _0x51c48c['kJeMt'](_0x205411,_0xf66968);},'uWLKw':function(_0x29e5f9,_0x5a123d){return _0x29e5f9===_0x5a123d;},'pscAA':_0xf668('2c6','%W95'),'gyEFZ':function(_0x9a6e05,_0x1e11e2){return _0x9a6e05(_0x1e11e2);}};if(_0x51c48c[_0xf668('2c7','wpRu')](_0x51c48c['KCdyN'],_0x51c48c[_0xf668('2c8','W6#w')])){$[_0xf668('2c9','lVb(')]({'url':_0xf668('2ca','DLvm')},(_0x509e8a,_0x452baa,_0x11f471)=>{var _0x5781eb={'VmgSP':function(_0x422d96){return _0x5be915[_0xf668('2cb','s^UP')](_0x422d96);}};try{if(_0x5be915[_0xf668('2cc','wpRu')](_0x5be915[_0xf668('2cd','%8LH')],_0x5be915['FuWzi'])){if(_0x509e8a){console['log'](''+JSON[_0xf668('2ce','LgTp')](_0x509e8a));}else{_0x11f471=JSON[_0xf668('2cf','CugW')](_0x11f471);if(_0x11f471[_0xf668('2d0','t1Hx')]){$[_0xf668('2d1','DqQY')]=_0x11f471['data'][_0xf668('2d2','yr@1')];}}}else{if(_0x509e8a){console[_0xf668('2d3','ERd%')](''+JSON[_0xf668('2d4','922l')](_0x509e8a));}else{if(_0x11f471){if(_0x5be915['xbxni'](_0x5be915[_0xf668('2d5','nKU7')],_0x5be915[_0xf668('2d6','vPrL')])){_0x11f471=JSON[_0xf668('2d7','^p1!')](_0x11f471);if(_0x5be915[_0xf668('2d8','On6g')](_0x11f471[_0xf668('2d9','wpRu')][_0xf668('2da',']AF0')],0x0)){$[_0xf668('2db','6BUo')]=_0x11f471['data'][0x0][_0xf668('2dc','zzNq')];}else{if(_0xf668('2dd','9qTx')===_0xf668('2de','DLvm')){console[_0xf668('2df','rHh*')]('获得1次翻牌机会');}else{$['userShareCode']='';}}}else{_0x3fa36f();}}}}}catch(_0x21a295){$[_0xf668('2e0','nKU7')](_0x21a295,_0x452baa);}finally{if(_0x5be915[_0xf668('2e1','KEOw')](_0xf668('2e2','On6g'),_0x5be915['pscAA'])){_0x5be915[_0xf668('2e3','kmgF')](_0x3fa36f,_0x11f471);}else{$[_0xf668('8c','[XoU')]=JSON[_0xf668('2e4','wpRu')](_0x11f471);_0x5781eb[_0xf668('2e5',']l!6')](_0x3fa36f);}}});}else{$['logErr'](err);}});}function submitShareCode(_0x225399){var _0x250371={'OXoDN':_0xf668('2e6','yr@1'),'lIjfw':'set-cookie','ttyPW':function(_0x45f724,_0x3facc1){return _0x45f724!==_0x3facc1;},'hNvvo':_0xf668('2e7','6BUo'),'ymiss':_0xf668('2e8','vPrL'),'LHpxE':function(_0x1658b9,_0x298380){return _0x1658b9(_0x298380);},'WYRnv':function(_0x5b9419,_0x3316da){return _0x5b9419===_0x3316da;},'VDBLg':_0xf668('2e9','x1Fm'),'hgDxJ':_0xf668('2ea','W6#w')};let _0x2646cc={'url':_0xf668('2eb','UDkl'),'headers':{'Content-Type':_0x250371[_0xf668('2ec','W6#w')]},'body':JSON[_0xf668('2ed','UDkl')](_0x225399)};return new Promise(async _0x5ed11e=>{var _0x1edc6c={'iQKOR':function(_0x30eddb,_0x46599f){return _0x250371[_0xf668('2ee',']l!6')](_0x30eddb,_0x46599f);},'RyABX':_0x250371[_0xf668('2ef','6BUo')],'xSpeq':'KztMN','EGiof':_0x250371[_0xf668('2f0','yzvA')],'rEBNR':function(_0x180843,_0x4d275b){return _0x250371[_0xf668('2f1','$eJr')](_0x180843,_0x4d275b);}};if(_0x250371[_0xf668('2f2','#IEW')](_0x250371['VDBLg'],_0xf668('2f3','UDkl'))){$[_0xf668('2f4','DqQY')](_0x2646cc,(_0x580456,_0x1eab5c,_0x340d95)=>{var _0x5b13bb={'ZlYfZ':function(_0x123b3d){return _0x123b3d();}};try{if(_0x1edc6c['iQKOR'](_0x1edc6c['RyABX'],_0xf668('2f5','BI4O'))){$[_0xf668('2f6','%W95')](_0x2646cc,(_0x3b9927,_0x1ded2d,_0x566944)=>{if(_0x566944){$['zRes']=JSON[_0xf668('2f7','DLvm')](_0x566944);_0x5b13bb[_0xf668('2f8','O3X^')](_0x5ed11e);};});}else{if(_0x580456){console[_0xf668('14d','yr@1')](''+JSON[_0xf668('2f9','dmO7')](_0x580456));console[_0xf668('2fa','On6g')]($['name']+_0xf668('2fb','[XoU'));}else{_0x340d95=JSON[_0xf668('2fc','yr@1')](_0x340d95);console[_0xf668('278','CugW')](_0x340d95[_0xf668('2fd','UDkl')]);}}}catch(_0x4804a0){if(_0x1edc6c['xSpeq']!==_0x1edc6c[_0xf668('2fe','@SZ@')]){$['logErr'](_0x4804a0,_0x1eab5c);}else{console[_0xf668('2ff','922l')](''+JSON[_0xf668('279','uhqA')](_0x580456));}}finally{_0x1edc6c[_0xf668('300','ERd%')](_0x5ed11e,_0x340d95);}});}else{cookie=resp[_0x250371[_0xf668('301','BI4O')]][_0x250371[_0xf668('302','WfWy')]]['join'](';')+';\x20'+originCookie;}});}function checkCookie(){var _0x187853={'XjQJx':function(_0x1ee366,_0x21ded5){return _0x1ee366===_0x21ded5;},'pSCJL':'1001','FkAdc':function(_0x4a136d,_0x297d53){return _0x4a136d===_0x297d53;},'Fnton':_0xf668('303','yr@1'),'cBZdd':'京东返回了空数据','Lwkyn':function(_0x508b7b,_0x4ec8ae){return _0x508b7b===_0x4ec8ae;},'xnzAB':_0xf668('304','nKU7'),'cQKpP':'WwMGq','BSGZm':function(_0x103a69){return _0x103a69();},'IkTBE':_0xf668('305','[XoU'),'BBkwe':_0xf668('306','t1Hx'),'JuURs':_0xf668('307','6BUo'),'aDUGo':_0xf668('308','%8LH')};const _0x3f48e0={'url':_0x187853[_0xf668('309','nKU7')],'headers':{'Host':_0xf668('30a','yr@1'),'Accept':_0x187853['BBkwe'],'Connection':'keep-alive','Cookie':cookie,'User-Agent':_0x187853['JuURs'],'Accept-Language':'zh-cn','Referer':_0xf668('30b','AA@b'),'Accept-Encoding':_0x187853[_0xf668('30c','kmgF')]}};return new Promise(_0x497d5e=>{$[_0xf668('30d','Ul$D')](_0x3f48e0,(_0x52be6b,_0x9525e7,_0x39b4fc)=>{try{if(_0x52be6b){$[_0xf668('30e','9qTx')](_0x52be6b);}else{if(_0x39b4fc){_0x39b4fc=JSON[_0xf668('7c','[XoU')](_0x39b4fc);if(_0x187853[_0xf668('30f','lVb(')](_0x39b4fc[_0xf668('310','W6#w')],_0x187853[_0xf668('311','rHh*')])){$[_0xf668('e','nKU7')]=![];return;}if(_0x187853[_0xf668('312','922l')](_0x39b4fc[_0xf668('313','yr@1')],'0')&&_0x39b4fc[_0xf668('22a','UDkl')]['hasOwnProperty'](_0x187853['Fnton'])){$['nickName']=_0x39b4fc[_0xf668('314','922l')][_0xf668('315','x1Fm')][_0xf668('316','[XoU')][_0xf668('317','9qTx')];}}else{$['log'](_0x187853[_0xf668('318','R[HN')]);}}}catch(_0x56d1d9){$[_0xf668('319','wpRu')](_0x56d1d9);}finally{if(_0x187853[_0xf668('31a','V[6Q')](_0x187853['xnzAB'],_0x187853[_0xf668('31b','[XoU')])){let _0x5b764e=newPosition['indexOf'](v['position']);newPosition['splice'](_0x5b764e,0x1);}else{_0x187853[_0xf668('31c','922l')](_0x497d5e);}}});});};_0xod3='jsjiami.com.v6';
-// prettier-ignore
-function Env(t,e){class s{constructor(t){this.env=t}send(t,e="GET"){t="string"==typeof t?{url:t}:t;let s=this.get;return"POST"===e&&(s=this.post),new Promise((e,i)=>{s.call(this,t,(t,s,r)=>{t?i(t):e(s)})})}get(t){return this.send.call(this.env,t)}post(t){return this.send.call(this.env,t,"POST")}}return new class{constructor(t,e){this.name=t,this.http=new s(this),this.data=null,this.dataFile="box.dat",this.logs=[],this.isMute=!1,this.isNeedRewrite=!1,this.logSeparator="\n",this.startTime=(new Date).getTime(),Object.assign(this,e),this.log("",`\ud83d\udd14${this.name}, \u5f00\u59cb!`)}isNode(){return"undefined"!=typeof module&&!!module.exports}isQuanX(){return"undefined"!=typeof $task}isSurge(){return"undefined"!=typeof $httpClient&&"undefined"==typeof $loon}isLoon(){return"undefined"!=typeof $loon}toObj(t,e=null){try{return JSON.parse(t)}catch{return e}}toStr(t,e=null){try{return JSON.stringify(t)}catch{return e}}getjson(t,e){let s=e;const i=this.getdata(t);if(i)try{s=JSON.parse(this.getdata(t))}catch{}return s}setjson(t,e){try{return this.setdata(JSON.stringify(t),e)}catch{return!1}}getScript(t){return new Promise(e=>{this.get({url:t},(t,s,i)=>e(i))})}runScript(t,e){return new Promise(s=>{let i=this.getdata("@chavy_boxjs_userCfgs.httpapi");i=i?i.replace(/\n/g,"").trim():i;let r=this.getdata("@chavy_boxjs_userCfgs.httpapi_timeout");r=r?1*r:20,r=e&&e.timeout?e.timeout:r;const[o,h]=i.split("@"),a={url:`http://${h}/v1/scripting/evaluate`,body:{script_text:t,mock_type:"cron",timeout:r},headers:{"X-Key":o,Accept:"*/*"}};this.post(a,(t,e,i)=>s(i))}).catch(t=>this.logErr(t))}loaddata(){if(!this.isNode())return{};{this.fs=this.fs?this.fs:require("fs"),this.path=this.path?this.path:require("path");const t=this.path.resolve(this.dataFile),e=this.path.resolve(process.cwd(),this.dataFile),s=this.fs.existsSync(t),i=!s&&this.fs.existsSync(e);if(!s&&!i)return{};{const i=s?t:e;try{return JSON.parse(this.fs.readFileSync(i))}catch(t){return{}}}}}writedata(){if(this.isNode()){this.fs=this.fs?this.fs:require("fs"),this.path=this.path?this.path:require("path");const t=this.path.resolve(this.dataFile),e=this.path.resolve(process.cwd(),this.dataFile),s=this.fs.existsSync(t),i=!s&&this.fs.existsSync(e),r=JSON.stringify(this.data);s?this.fs.writeFileSync(t,r):i?this.fs.writeFileSync(e,r):this.fs.writeFileSync(t,r)}}lodash_get(t,e,s){const i=e.replace(/\[(\d+)\]/g,".$1").split(".");let r=t;for(const t of i)if(r=Object(r)[t],void 0===r)return s;return r}lodash_set(t,e,s){return Object(t)!==t?t:(Array.isArray(e)||(e=e.toString().match(/[^.[\]]+/g)||[]),e.slice(0,-1).reduce((t,s,i)=>Object(t[s])===t[s]?t[s]:t[s]=Math.abs(e[i+1])>>0==+e[i+1]?[]:{},t)[e[e.length-1]]=s,t)}getdata(t){let e=this.getval(t);if(/^@/.test(t)){const[,s,i]=/^@(.*?)\.(.*?)$/.exec(t),r=s?this.getval(s):"";if(r)try{const t=JSON.parse(r);e=t?this.lodash_get(t,i,""):e}catch(t){e=""}}return e}setdata(t,e){let s=!1;if(/^@/.test(e)){const[,i,r]=/^@(.*?)\.(.*?)$/.exec(e),o=this.getval(i),h=i?"null"===o?null:o||"{}":"{}";try{const e=JSON.parse(h);this.lodash_set(e,r,t),s=this.setval(JSON.stringify(e),i)}catch(e){const o={};this.lodash_set(o,r,t),s=this.setval(JSON.stringify(o),i)}}else s=this.setval(t,e);return s}getval(t){return this.isSurge()||this.isLoon()?$persistentStore.read(t):this.isQuanX()?$prefs.valueForKey(t):this.isNode()?(this.data=this.loaddata(),this.data[t]):this.data&&this.data[t]||null}setval(t,e){return this.isSurge()||this.isLoon()?$persistentStore.write(t,e):this.isQuanX()?$prefs.setValueForKey(t,e):this.isNode()?(this.data=this.loaddata(),this.data[e]=t,this.writedata(),!0):this.data&&this.data[e]||null}initGotEnv(t){this.got=this.got?this.got:require("got"),this.cktough=this.cktough?this.cktough:require("tough-cookie"),this.ckjar=this.ckjar?this.ckjar:new this.cktough.CookieJar,t&&(t.headers=t.headers?t.headers:{},void 0===t.headers.Cookie&&void 0===t.cookieJar&&(t.cookieJar=this.ckjar))}get(t,e=(()=>{})){t.headers&&(delete t.headers["Content-Type"],delete t.headers["Content-Length"]),this.isSurge()||this.isLoon()?(this.isSurge()&&this.isNeedRewrite&&(t.headers=t.headers||{},Object.assign(t.headers,{"X-Surge-Skip-Scripting":!1})),$httpClient.get(t,(t,s,i)=>{!t&&s&&(s.body=i,s.statusCode=s.status),e(t,s,i)})):this.isQuanX()?(this.isNeedRewrite&&(t.opts=t.opts||{},Object.assign(t.opts,{hints:!1})),$task.fetch(t).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>e(t))):this.isNode()&&(this.initGotEnv(t),this.got(t).on("redirect",(t,e)=>{try{if(t.headers["set-cookie"]){const s=t.headers["set-cookie"].map(this.cktough.Cookie.parse).toString();s&&this.ckjar.setCookieSync(s,null),e.cookieJar=this.ckjar}}catch(t){this.logErr(t)}}).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>{const{message:s,response:i}=t;e(s,i,i&&i.body)}))}post(t,e=(()=>{})){if(t.body&&t.headers&&!t.headers["Content-Type"]&&(t.headers["Content-Type"]="application/x-www-form-urlencoded"),t.headers&&delete t.headers["Content-Length"],this.isSurge()||this.isLoon())this.isSurge()&&this.isNeedRewrite&&(t.headers=t.headers||{},Object.assign(t.headers,{"X-Surge-Skip-Scripting":!1})),$httpClient.post(t,(t,s,i)=>{!t&&s&&(s.body=i,s.statusCode=s.status),e(t,s,i)});else if(this.isQuanX())t.method="POST",this.isNeedRewrite&&(t.opts=t.opts||{},Object.assign(t.opts,{hints:!1})),$task.fetch(t).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>e(t));else if(this.isNode()){this.initGotEnv(t);const{url:s,...i}=t;this.got.post(s,i).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>{const{message:s,response:i}=t;e(s,i,i&&i.body)})}}time(t){let e={"M+":(new Date).getMonth()+1,"d+":(new Date).getDate(),"H+":(new Date).getHours(),"m+":(new Date).getMinutes(),"s+":(new Date).getSeconds(),"q+":Math.floor(((new Date).getMonth()+3)/3),S:(new Date).getMilliseconds()};/(y+)/.test(t)&&(t=t.replace(RegExp.$1,((new Date).getFullYear()+"").substr(4-RegExp.$1.length)));for(let s in e)new RegExp("("+s+")").test(t)&&(t=t.replace(RegExp.$1,1==RegExp.$1.length?e[s]:("00"+e[s]).substr((""+e[s]).length)));return t}msg(e=t,s="",i="",r){const o=t=>{if(!t)return t;if("string"==typeof t)return this.isLoon()?t:this.isQuanX()?{"open-url":t}:this.isSurge()?{url:t}:void 0;if("object"==typeof t){if(this.isLoon()){let e=t.openUrl||t.url||t["open-url"],s=t.mediaUrl||t["media-url"];return{openUrl:e,mediaUrl:s}}if(this.isQuanX()){let e=t["open-url"]||t.url||t.openUrl,s=t["media-url"]||t.mediaUrl;return{"open-url":e,"media-url":s}}if(this.isSurge()){let e=t.url||t.openUrl||t["open-url"];return{url:e}}}};if(this.isMute||(this.isSurge()||this.isLoon()?$notification.post(e,s,i,o(r)):this.isQuanX()&&$notify(e,s,i,o(r))),!this.isMuteLog){let t=["","==============\ud83d\udce3\u7cfb\u7edf\u901a\u77e5\ud83d\udce3=============="];t.push(e),s&&t.push(s),i&&t.push(i),console.log(t.join("\n")),this.logs=this.logs.concat(t)}}log(...t){t.length>0&&(this.logs=[...this.logs,...t]),console.log(t.join(this.logSeparator))}logErr(t,e){const s=!this.isSurge()&&!this.isQuanX()&&!this.isLoon();s?this.log("",`\u2757\ufe0f${this.name}, \u9519\u8bef!`,t.stack):this.log("",`\u2757\ufe0f${this.name}, \u9519\u8bef!`,t)}wait(t){return new Promise(e=>setTimeout(e,t))}done(t={}){const e=(new Date).getTime(),s=(e-this.startTime)/1e3;this.log("",`\ud83d\udd14${this.name}, \u7ed3\u675f! \ud83d\udd5b ${s} \u79d2`),this.log(),(this.isSurge()||this.isQuanX()||this.isLoon())&&$done(t)}}(t,e)}
